% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.tables.R
\name{add.tables}
\alias{add.tables}
\title{Add stat table to an existing contingency table}
\usage{
add.tables(x, format = NULL, ...)
}
\arguments{
\item{x}{A crosstable object}

\item{format}{Character vector with the format of the stats. See help("\%f\%")}

\item{...}{A list of tables. They must have the same dimensions of the original table.}
}
\description{
Add stat table to an existing contingency table
}
\examples{
library(magrittr)

bacteria <- MASS::bacteria

tab <- xtabs(~y+trt, data=bacteria)
tab_chsq <- chisq.test(tab)

crosstable(tab) \%>\%
  add.tables(format=c("0.00\%", "(0.00)"), "\%" = prop.table(tab)*100, "expected" = tab_chsq$expected)

}
