% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{add_header_list}
\alias{add_header_list}
\title{add_header_list}
\usage{
add_header_list(x, top = TRUE, args)
}
\arguments{
\item{x}{x}

\item{top}{top}

\item{args}{args}
}
\description{
add_header_list
}
\author{
David Hajage
}
\keyword{internal}
