% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{addAlert}
\alias{addAlert}
\title{Add a new paragraph with a Alert style}
\usage{
addAlert(doc, value, stylename = "Alert", ...)
}
\arguments{
\item{doc}{the doc object (created with the docx function of ReporteRs package)}

\item{value}{a character string}

\item{stylename}{a character string}

\item{...}{...}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Alert style
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- addAlert(doc, "Coucou")
}
}
\author{
David Hajage
}
\keyword{univar}
