% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{addCode}
\alias{addCode}
\title{Add a new paragraph with a Code style}
\usage{
addCode(doc, value, ...)
}
\arguments{
\item{doc}{the doc object (created with the docx function of ReporteRs package)}

\item{value}{a character string}

\item{...}{...}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Code style
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- addCode(doc, "Coucou")
}
}
\author{
David Hajage
}
\keyword{univar}
