% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{addCrossTable}
\alias{addCrossTable}
\title{add a table made by the cross function into a ReporteRs document}
\usage{
addCrossTable(doc, crosstable, compact = FALSE)
}
\arguments{
\item{doc}{a \code{docx} object created by \code{docx} function (package \code{ReporteRs})}

\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}
}
\value{
A \code{docx} object
}
\description{
add a table made by the cross function into a ReporteRs document
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
doc <- docx()
doc <- addCrossTable(doc, mytable)
doc <- addPageBreak(doc)
doc <- addCrossTable(doc, mytable, TRUE)
}
}
\author{
David Hajage
}
\keyword{univar}

