% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{addEnumerate}
\alias{addEnumerate}
\title{Add a list of numerated items}
\usage{
addEnumerate(doc, value, level = 1, ...)
}
\arguments{
\item{doc}{the doc object (created with the docx function of ReporteRs package)}

\item{value}{a vector of character strings}

\item{level}{a integer (1 to 4)}
}
\value{
a new doc object
}
\description{
Add a list of numerated items
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- addEnumerate(doc, c("Coucou", "Cuicui"))
}
}
\author{
David Hajage
}
\keyword{univar}
