% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{addVerbatim}
\alias{addVerbatim}
\title{Add a new paragraph with a Verbatim style}
\usage{
addVerbatim(doc, value, stylename = "Verbatim", ...)
}
\arguments{
\item{doc}{the doc object (created with the docx function of ReporteRs package)}

\item{value}{a character string}

\item{stylename}{a character string}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Verbatim style
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- addVerbatim(doc, "Coucou")
}
}
\author{
David Hajage
}
\keyword{univar}
