% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{apply_labels}
\alias{apply_labels}
\title{Batch set variable labels}
\usage{
apply_labels(data, ..., warn_missing = FALSE)
}
\arguments{
\item{data}{data.frame/list}

\item{...}{named arguments}
}
\description{
This function is a copycat of from expss package v0.10.7 (slightly modified) to avoid having to depend on expss. See \code{\link[expss:apply_labels]{expss::apply_labels()}} for more documentation. Note that this version is not compatible with \code{data.table}.
}
\examples{
library(crosstable)
cars \%>\%
  apply_labels(speed="Speed (mph)",
             dist="Stopping distance (ft)") \%>\% 
  crosstable()
}
