% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{as.crosstable}
\alias{as.crosstable}
\title{Coerce to a Crosstable (for officer docx addition)}
\usage{
as.crosstable(df, labs.col = "???")
}
\arguments{
\item{df}{a data.frame}

\item{labs.col}{the name of the grouping variable}
}
\value{
a cross
}
\description{
Coerce to a Crosstable (for officer docx addition)
}
\examples{
library(dplyr) #for the pipe operator
library(officer)
mytable = cross(cbind(Sepal.Length, I(Sepal.Width^2)) ~ Species, iris) \%>\% 
   as.data.frame \%>\% #loses attributes
   as.crosstable(labs.col = "Species")
   
doc <- read_docx() \%>\% 
    body_add_crosstable(mytable) 

\dontrun{
dfile <- "test_doc.docx"
print(doc, target = dfile)
shell.exec(dfile)
}

}
