% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{as_gt.crosstable}
\alias{as_gt.crosstable}
\title{Converts a \code{crosstable} object into a formatted \code{gt} table.}
\usage{
\method{as_gt}{crosstable}(
  x,
  show_test_name = TRUE,
  by_header = NULL,
  keep_id = FALSE,
  generic_labels = list(id = ".id", variable = "variable", value = "value", total =
    "Total", label = "label", test = "test", effect = "effect"),
  ...
)
}
\arguments{
\item{x}{the result of \code{\link[=crosstable]{crosstable()}}}

\item{show_test_name}{in the \code{test} column, show the test name}

\item{by_header}{a string to override the \code{by} header}

\item{keep_id}{whether to keep the \code{.id} column}

\item{generic_labels}{names of the crosstable default columns}

\item{...}{unused}
}
\description{
Converts a \code{crosstable} object into a formatted \code{gt} table.
}
\examples{
xx = mtcars2 \%>\% dplyr::select(1:9)
crosstable(xx) \%>\% as_gt
crosstable(xx, by=am) \%>\% as_gt
crosstable(xx, by=am, test=TRUE, total=TRUE, effect=TRUE) \%>\% as_gt(keep_id=TRUE)
}
\seealso{
\code{\link[=as_flextable.crosstable]{as_flextable.crosstable()}}

Other as_gt methods: 
\code{\link{as_gt.default}()},
\code{\link{as_gt}()}
}
\author{
Dan Chaltiel
}
\concept{as_gt methods}
