% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{as_gt.default}
\alias{as_gt.default}
\title{Default method to convert an object to a \code{gt} table}
\usage{
\method{as_gt}{default}(x, ...)
}
\arguments{
\item{x}{object to be converted}

\item{...}{arguments for custom methods}
}
\description{
Default method to convert an object to a \code{gt} table
}
\seealso{
Other as_gt methods: 
\code{\link{as_gt.crosstable}()},
\code{\link{as_gt}()}
}
\concept{as_gt methods}
