% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{as_gt.crosstable}
\alias{as_gt.crosstable}
\alias{as_gt}
\title{Turns a \code{crosstable} object into a formatted \code{gt} table}
\usage{
\method{as_gt}{crosstable}(
  data,
  show_test_name = TRUE,
  by_header = NULL,
  keep_id = FALSE,
  generic_labels = list(id = ".id", variable = "variable", value = "value", total =
    "Total", label = "label", test = "test", effect = "effect"),
  ...
)

as_gt(data, ...)
}
\arguments{
\item{data}{}

\item{show_test_name}{in the \code{test} column, show the test name}

\item{by_header}{a string to override the \code{by} header}

\item{keep_id}{whether to keep the \code{.id} column}

\item{generic_labels}{names of the crosstable default columns}

\item{...}{}
}
\description{
Turns a \code{crosstable} object into a formatted \code{gt} table

Generic method
}
\section{Methods (by class)}{
\itemize{
\item \code{crosstable}: Turns a \code{crosstable} object into a formatted \code{gt} table.
}}

\examples{
xx = mtcars2 \%>\% select(1:9)
crosstable(xx) \%>\% as_gt
crosstable(xx, by=am) \%>\% as_gt
crosstable(xx, by=am, test=T, total=T, effect=T) \%>\% as_gt(keep_id=TRUE)
}
\seealso{
\code{\link[=as_flextable.crosstable]{as_flextable.crosstable()}}

\code{\link[gt:gt]{gt::gt()}}
}
\author{
Dan Chaltiel
}
\concept{as_gt methods}
