% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.tables.R
\name{bind.tables}
\alias{bind.tables}
\title{Join two tables}
\usage{
bind.tables(...)
}
\description{
Function borrwed from RKward to bind tables.
You can use it to join two statistic and format them with ftable
}
\examples{
tab1 <- xtabs(Freq ~ Type + Cont, data=MASS::housing)
ptab1 <- prop.table(tab1)

# Join the a contingency table with the proportion table
bind.tables(n = tab1, prop = ptab1)

# You can use ftable to get a more readable output
ftable(bind.tables(n = tab1, prop = ptab1))

}
