% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{biostats2SummaryFunctions}
\alias{biostats2SummaryFunctions}
\alias{moystd}
\alias{mediqr}
\alias{minmax}
\alias{nna}
\title{Summary functions}
\usage{
moystd(
  x,
  na.rm = TRUE,
  dig = 2,
  fixed_format = getOption("crosstable_fixed_format", FALSE),
  zero_digits = 1
)

mediqr(
  x,
  na.rm = TRUE,
  dig = 2,
  fixed_format = getOption("crosstable_fixed_format", FALSE),
  zero_digits = 1
)

minmax(x, na.rm = TRUE, dig = 2)

nna(x)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{\code{TRUE} as default}

\item{dig}{number of digits}

\item{fixed_format}{use \code{\link{format_fixed}} instead of round (default)}

\item{zero_digits}{is fixed_format==TRUE, number of significant digits for values rounded to 0 (set to NULL to keep the original 0 value)}
}
\description{
Summary functions to use with \code{\link{cross}} or anywhere else.
}
\section{Functions}{
\itemize{
\item \code{moystd}: returns mean and std error

\item \code{mediqr}: returns median and IQR

\item \code{minmax}: returns minimum and maximum

\item \code{nna}: returns  number of observations and number of missing values
}}

\section{Fixed format}{

The use of the \code{fixed_format} argument allows to have trailing zeros after rounded values.
In the case when the output of rounded values is zero, the use of the \code{zero_digits} argument allows to keep some significant digits for this specific case only.
}

\examples{
moystd(iris$Sepal.Length, dig=3)
minmax(iris$Sepal.Length, dig=3)
mediqr(iris$Sepal.Length, dig=3)
nna(iris$Sepal.Length)

x = iris$Sepal.Length/10000 #closer to zero 

moystd(x, dig=3, fixed_format=TRUE)
moystd(x, dig=3, fixed_format=TRUE, zero_digits=NULL)
options("crosstable_fixed_format"=TRUE)
moystd(x, dig=3, zero_digits=2)
options("crosstable_fixed_format"=NULL)

}
\seealso{
\code{\link{format_fixed}}
}
\author{
Dan Chaltiel

David Hajage
}
