% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_flextable2}
\alias{body_add_flextable2}
\title{Alternative to \code{\link[flextable:body_add_flextable]{flextable::body_add_flextable()}}}
\usage{
body_add_flextable2(doc, x, bookmark = NULL, append_line = TRUE, ...)
}
\arguments{
\item{doc}{An \code{rdocx} object from \strong{officer}.}

\item{x}{A \code{flextable} object. If it has a \code{"legend"} attribute, it is added.}

\item{bookmark}{Optional. Word bookmark name for the legend.}

\item{append_line}{Whether to add an empty line after the table.}

\item{...}{Passed to \code{\link[flextable:body_add_flextable]{flextable::body_add_flextable()}}.}
}
\value{
The docx object \code{doc}
}
\description{
Extends \code{body_add_flextable()} by adding:
\itemize{
\item a legend (if \code{x} has a \code{"legend"} attribute), via \code{\link[=body_add_table_legend]{body_add_table_legend()}}
\item an optional empty line after the table
}
}
\examples{
library(officer)
library(ggplot2)
ft = flextable::flextable(head(iris)) \%>\%
  structure(legend="The iris dataset")

doc = read_docx() \%>\%
 body_add_normal("Text before") \%>\%
 body_add_flextable2(ft) \%>\%
 body_add_normal("Text after")
write_and_open(doc)
}
