% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_gg2}
\alias{body_add_gg2}
\title{Alternative to \code{\link[officer:body_add_gg]{officer::body_add_gg()}} which uses \code{ggplot} syntax}
\usage{
body_add_gg2(
  doc,
  value,
  width = 6,
  height = 5,
  units = c("in", "cm", "mm"),
  res = 300,
  style = getOption("crosstable_style_image", doc$default_styles$paragraph),
  ...
)
}
\arguments{
\item{doc}{an \code{rdocx} object}

\item{value}{ggplot object}

\item{width, height}{width and height. Can be abbreviated.}

\item{units}{}

\item{res}{resolution of the png image in ppi (passed to the argument \code{dpi} of \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}})}

\item{style}{paragraph style}

\item{...}{other arguments to be passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
Alternative to \code{\link[officer:body_add_gg]{officer::body_add_gg()}} which uses \code{ggplot} syntax
}
\examples{
\dontrun{
if(require("ggplot2") && capabilities(what = "png")){
  library(officer)
  p = ggplot(data = iris ) +
    geom_point(mapping = aes(Sepal.Length, Petal.Length))
  options(crosstable_style_image="centered")
  doc = read_docx() \%>\%
    body_add_normal("Text before") \%>\% 
    body_add_gg2(p, w=14, h=10, units="cm") \%>\%
    body_add_normal("Text after")
  write_and_open(doc, "example.docx")
}
}
}
