% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_glued}
\alias{body_add_glued}
\title{Add a new paragraph with a Normal style to an \code{officer} document, inserting variables with \code{glue::glue}}
\usage{
body_add_glued(doc, .x, ...)
}
\arguments{
\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}

\item{.x}{the string with \code{glue::glue} patterns (Expressions enclosed by braces will be evaluated as R code)}

\item{...}{parameters to be passed to \code{glue::glue}}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Normal style to an \code{officer} document, inserting variables with \code{glue::glue}
}
\examples{
\dontrun{
library(officer)
library(crosstable)
library(dplyr)
doc = read_docx()
doc = doc \%>\%
    body_add_glued("The iris table has {ncol(iris)} columns") \%>\%
    body_add_glued("Actually, it also has {nrows} rows", nrows=nrow(iris))
}
}
\author{
Dan Chaltiel
}
