% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_img2}
\alias{body_add_img2}
\title{Alternative to \code{\link[officer:body_add_img]{officer::body_add_img()}} which adds a \code{units} choice}
\usage{
body_add_img2(doc, src, width, height, units = c("in", "cm", "mm"), ...)
}
\arguments{
\item{doc}{an \code{rdocx} object}

\item{src}{image filename, the basename of the file must not contain any blank.}

\item{width, height}{width and height}

\item{units}{units for width and height}

\item{...}{other arguments to be passed to \code{\link[officer:body_add_img]{officer::body_add_img()}}}
}
\description{
Alternative to \code{\link[officer:body_add_img]{officer::body_add_img()}} which adds a \code{units} choice
}
