% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_normal}
\alias{body_add_normal}
\title{Add a new paragraph with a Normal style to an \code{officer} document, inserting variables with \code{base::paste}}
\usage{
body_add_normal(doc, ...)
}
\arguments{
\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}

\item{...}{one or several character strings, collapsed into a paragraph with \code{base::paste}}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Normal style to an \code{officer} document, inserting variables with \code{base::paste}
}
\examples{
\dontrun{
library(officer)
library(crosstable)
library(dplyr)
doc = read_docx()
doc = doc \%>\% body_add_normal("La table iris a ", ncol(iris), " colonnes.")
}
}
\author{
Dan Chaltiel
}
