% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_table_legend}
\alias{body_add_table_legend}
\alias{body_add_figure_legend}
\title{Add a table legend to an \code{officer} document}
\usage{
body_add_table_legend(
  doc,
  legend,
  bookmark = NULL,
  legend_style = getOption("crosstable_style_legend", "Table Caption"),
  style = getOption("crosstable_style_strong", "strong"),
  seqfield = "SEQ Table \\\\* Arabic"
)

body_add_figure_legend(
  doc,
  legend,
  bookmark = NULL,
  legend_style = getOption("crosstable_style_legend", "Image Caption"),
  style = getOption("crosstable_style_strong", "strong"),
  seqfield = "SEQ Figure \\\\* Arabic"
)
}
\arguments{
\item{doc}{a docx object}

\item{legend}{the table legend}

\item{legend_style}{may depend on the docx template}

\item{style}{the legend style (strong, italic...)}

\item{seqfield}{to figure this out, in a docx file, insert a table legend, right click on the inserted number and select "Toggle Field Codes". This argument should be the value of the field, with extra escaping.}
}
\description{
Add a table legend to an \code{officer} document
}
\section{Warning}{

At first, the legends added with \code{\link[=body_add_table_legend]{body_add_table_legend()}} or \code{\link[=body_add_figure_legend]{body_add_figure_legend()}} have no numbers. You have to manualy update the references in MS Word: select all (\kbd{Ctrl}+\kbd{A}), then update (\kbd{F9}). You might have to do this several times. More info on \link{https://ardata-fr.github.io/officeverse/faq.html#update-fields}.
}

\examples{
\dontrun{
library(officer)
library(ggplot2)
p=quickplot(x=Sepal.Length, y=Sepal.Width, color=Species, data=iris)
x=read_docx() \%>\% 
  body_add_normal("As you can see in Table \\\\@ref(tab1) and in Figure \\\\@ref(fig1), the iris dataset is about flowers.") \%>\% 
  body_add_normal() \%>\% 
  body_add_table_legend("Iris dataset", bookmark="tab1") \%>\% 
  body_add_crosstable(crosstable(iris)) \%>\% 
  body_add_gg(p) \%>\% 
  body_add_figure_legend("Iris plot", bookmark="fig1")
write_and_open(x, "example.docx")
#press Ctrl+A then F9 twice for the reference to appear.
}
}
\author{
Dan Chaltiel
}
