% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{cbind.list}
\alias{cbind.list}
\title{Cbind for list}
\usage{
\method{cbind}{list}(l, preserve.attr = FALSE)
}
\arguments{
\item{l}{a list}
}
\description{
Cbind for list
}
\author{
David Hajage
}
\keyword{internal}
