% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{compact}
\alias{compact}
\title{Compact the result of cross function}
\usage{
compact(x)
}
\arguments{
\item{x}{a crosstable, the result of \code{cross} function}
}
\description{
Compact the result of cross function
}
\examples{
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
compact(mytable)
}
\author{
David Hajage
}
