% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint_numeric}
\alias{confint_numeric}
\title{Confidence interval of a numeric vector}
\usage{
confint_numeric(object, level = 0.95, B = 0)
}
\arguments{
\item{object}{a vector, numeric or equivalent (date, logical...)}

\item{level}{the confidence level required}

\item{B}{if >0, the number of bootstraps}
}
\value{
the vector [conf_inf, conf_sup]
}
\description{
Not an S3 method, which might have conflicted with \link[stats:confint]{stats::confint}.
}
\examples{
confint_numeric(iris$Sepal.Length)
confint_numeric(mtcars2$hp_date)
confint_numeric(mtcars2$hp_date, level=0.99)
}
\author{
Dan Chaltiel
}
