% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.r
\name{correlation.data.frame}
\alias{correlation.data.frame}
\title{Compute correlation (data.frame input)}
\usage{
correlation.data.frame(dfx, dfy, method = c("pearson", "kendall", "spearman"),
  digits = 2, label = FALSE)
}
\arguments{
\item{dfx}{data.frame}

\item{dfy}{data.frame}

\item{method}{method}

\item{digits}{digits}

\item{label}{label}
}
\description{
Compute correlation (data.frame input)
}
\author{
David Hajage
}
\keyword{internal}
