% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.r
\name{correlation}
\alias{correlation}
\title{Compute correlation}
\usage{
correlation(x, y, method = c("pearson", "kendall", "spearman"),
  digits = 2)
}
\arguments{
\item{x}{numerical object}

\item{y}{numerical object}

\item{method}{method}

\item{digits}{digits}
}
\description{
Compute correlation
}
\author{
David Hajage
}
\keyword{internal}
