% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{create.report}
\alias{create.report}
\title{Create a docx object with a specific template and informations already inserted}
\usage{
create.report(template = c("gerc", "urc", "cephepi"), title = "",
  acronym = "", version = "", npromo = "", nct = "", invest = "",
  biostat = "", methodo = "", date_lastmodif = "", date_freez = "",
  date_update = "", history = NULL)
}
\arguments{
\item{template}{either 'gerc', 'urc', or 'cephepi',}

\item{title}{Study title (character string)}

\item{acronym}{Study acronym (character string)}

\item{version}{Version number (character string)}

\item{npromo}{Promotion number (character string, not used with gerc template)}

\item{nct}{Clinical trial number (character string, not used with gerc template)}

\item{invest}{Name of the principal investigator (character string)}

\item{biostat}{Name of the biostatistician (character string)}

\item{methodo}{Name of the methodologist (character string, not used with gerc template)}

\item{date_lastmodif}{Date of last modification of the document (character string)}

\item{date_freez}{Date of data freezing (character string)}

\item{date_update}{Date of last history description update  (character string)}

\item{history}{History description. A list of length 1 (min) to 7 (max), each element being a list of four elements: list(version = "Version number", author = "Author", description = c("Description 1", "Description 2"), date = "Date")}
}
\value{
a FlexTable object
}
\description{
Create a docx object with a specific template and informations already inserted
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- create.report(template = 'urc', 
                     title = 'A great study', 
                     acronym = 'GREAT', 
                     version = '1.3', 
                     npromo = 'AOR17123', 
                     nct = 'NCT123456', 
                     invest = 'Casimir', 
                     biostat = 'Goldorak', 
                     methodo = 'Actarus', 
                     date_lastmodif = '23/03/1979', 
                     date_freez = '22/03/1979', 
                     date_update = '24/03/1979', 
                     history = list(
                                    list(version = '1', 
                                         author = 'Goldorak', 
                                         description = c('First description', 'Primary outcome analysis'), 
                                         date = '28/03/1979'), 
                                    list(version = '1.3', 
                                         author = 'Goldorak', 
                                         description = 'Add forgotten things', 
                                         date = '29/03/1979')
                                   )
                    )
}
}
\author{
David Hajage
}
\keyword{univar}
