% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix.r
\name{cross_all}
\alias{cross_all}
\title{Cross all x with all y}
\usage{
cross_all(x, y = NULL, funs = c(mean, sd, quantile, n, na), ...,
  margin = 0:2, total = FALSE, digits = 2, showNA = c("no", "ifany",
  "always"), method = c("pearson", "kendall", "spearman"), times = NULL,
  followup = FALSE, test = FALSE, test.tabular = test.tabular.auto,
  test.summarize = test.summarize.auto,
  test.survival = test.survival.logrank, show.test = display.test,
  plim = 4, show.method = TRUE, effect = FALSE,
  effect.summarize = diff.mean.auto, effect.tabular = or.row.by.col,
  effect.survival = effect.survival.coxph, conf.level = 0.95,
  label = FALSE)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{funs}{funs}

\item{...}{\dots}

\item{margin}{margin}

\item{total}{total}

\item{digits}{digits}

\item{showNA}{showNA}

\item{method}{method}

\item{times}{times}

\item{followup}{followup}

\item{test}{test}

\item{test.tabular}{test.tabular}

\item{test.summarize}{test.summarize}

\item{test.survival}{test.survival}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{effect}{effect}

\item{effect.summarize}{effect.summarize}

\item{effect.tabular}{effect.tabular}

\item{effect.survival}{effect.survival}

\item{conf.level}{conf.level}

\item{label}{label}
}
\description{
Cross all x with all y
}
\author{
David Hajage
}
\keyword{internal}
