% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remix.r
\name{cross_list}
\alias{cross_list}
\title{Cross variables in a list}
\usage{
cross_list(l, funs = c(mean, sd, quantile, n, na), ..., margin = 0:2,
  total = FALSE, digits = 2, showNA = c("no", "ifany", "always"),
  method = c("pearson", "kendall", "spearman"), times = NULL,
  followup = FALSE, test = FALSE, test.summarize = test.summarize.auto,
  test.survival = test.survival.logrank, test.tabular = test.tabular.auto,
  show.test = display.test, plim = 4, show.method = TRUE, label = FALSE)
}
\arguments{
\item{l}{}

\item{funs}{funs}

\item{...}{\dots}

\item{margin}{margin}

\item{total}{total}

\item{digits}{digits}

\item{showNA}{showNA}

\item{method}{method}

\item{times}{times}

\item{followup}{followup}

\item{test}{test}

\item{test.summarize}{test.summarize}

\item{test.survival}{}

\item{test.tabular}{test.tabular}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{label}{label}
}
\description{
Cross variables in a list
}
\author{
David Hajage
}
\keyword{internal}

