% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix.r
\name{cross_one}
\alias{cross_one}
\title{Cross one x and one y}
\usage{
cross_one(x, y = NULL, funs = c(mean, sd, quantile, n, na), ...,
  margin = 0:2, total = FALSE, digits = 2, showNA = c("no", "ifany",
  "always"), method = c("pearson", "kendall", "spearman"), times = NULL,
  followup = FALSE, test = FALSE, test.tabular = test.tabular.auto,
  test.summarize = test.summarize.auto,
  test.survival = test.survival.logrank, show.test = display.test,
  plim = 4, show.method = TRUE, label = FALSE)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{funs}{funs}

\item{...}{\dots}

\item{margin}{margin}

\item{total}{total}

\item{digits}{digits}

\item{showNA}{showNA}

\item{method}{method}

\item{times}{times}

\item{followup}{followup}

\item{test}{test}

\item{test.tabular}{test.tabular}

\item{test.summarize}{test.summarize}

\item{test.survival}{test.survival}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{label}{label}
}
\description{
Cross one x and one y
}
\author{
David Hajage
}
\keyword{internal}
