% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.r
\name{cross_summary}
\alias{cross_summary}
\alias{mysummary}
\title{Summarize a numeric vector}
\usage{
cross_summary(x, dig = 2)

mysummary(x, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{dig}{number of digits}
}
\description{
Summarize a numeric vector
}
\section{Note}{
 
Function \code{mysummary} is kept for compatibility with old codes. It produces the same exact object than \code{cross_summary}, which should be preferred.
}

\keyword{internal}
