% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_to_flextable.R, R/officer.R
\name{cross_to_flextable}
\alias{cross_to_flextable}
\alias{ctf}
\alias{body_add_crosstable}
\title{Crosstables output}
\usage{
cross_to_flextable(
  crosstable,
  auto.fit = TRUE,
  compact = FALSE,
  show.test.name = TRUE,
  by_header = NULL,
  keep_id = FALSE,
  generic.labels = list(id = ".id", variable = "variable", value = "value", total =
    "Total", label = "label", test = "test", effect = "effect")
)

body_add_crosstable(doc, ...)
}
\arguments{
\item{crosstable}{the result of \code{crosstable} function}

\item{auto.fit}{whether to \code{flextable::autofit} the table}

\item{compact}{whether to compact the table}

\item{show.test.name}{in the \code{test} column, show the test name}

\item{by_header}{a string to override the \code{by} header}

\item{keep_id}{whether to keep the \code{.id} column}

\item{generic.labels}{names of the crosstable default columns}

\item{doc}{a \code{rdocx} object created by \code{read_docx} function (see \code{officer} package)}

\item{...}{arguments for \code{cross_to_flextable}}
}
\description{
\code{\link[=cross_to_flextable]{cross_to_flextable()}} turns the result of \code{\link[=crosstable]{crosstable()}} into a formatted \code{flextable} object. \cr
You can also use the shortcut \code{ctf()} which has the same arguments.

\code{\link[=body_add_crosstable]{body_add_crosstable()}} adds such a \code{flextable} an \code{officer} document.
}
\examples{
library(dplyr)
crosstable(mtcars2) \%>\% cross_to_flextable()
crosstable(mtcars2, by=vs, test = TRUE) \%>\% cross_to_flextable()
crosstable(esoph, by=tobgp, test = TRUE) \%>\% cross_to_flextable()
crosstable(iris, by=Species, test = TRUE) \%>\% cross_to_flextable(auto.fit=FALSE)
crosstable(iris, by=Species, test = TRUE) \%>\% cross_to_flextable(compact=TRUE)
crosstable(iris) \%>\% cross_to_flextable(compact=TRUE, auto.fit=TRUE)

#Renaming
crosstable(iris, by=Species, total="both", test=TRUE, effect=TRUE) \%>\%
   rename(ID=.id, math=variable, Tot=Total, lab=label, pval=test, fx=effect) \%>\%
   cross_to_flextable(by_header = "The specie", 
                      generic.labels=list(id = "ID", variable = "math", total="Tot", 
                                          label = "lab", test = "pval", effect="fx"))
#Officer
library(officer)
mytable = crosstable(mtcars2)
doc = read_docx() \%>\% 
    body_add_crosstable(mytable) \%>\% 
    body_add_break \%>\% 
    body_add_crosstable(mytable, compact=TRUE)

\dontrun{
dfile = "examples\\\\example_doc.docx"
print(doc, target = dfile)
shell.exec(dfile)
}
}
\seealso{
\code{\link[=crosstable]{crosstable()}}, \code{\link[flextable:flextable]{flextable::flextable()}}
}
\author{
Dan Chaltiel
}
