% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable-package.R
\docType{package}
\name{crosstable-package}
\alias{crosstable-package}
\alias{_PACKAGE}
\alias{Test-package}
\title{crosstable: Crosstables for descriptive analyses}
\description{
A handy tool for description and testing.
}
\details{
The \code{crosstable}package is centered on the \code{\link[=crosstable]{crosstable()}}` function.
See vignettes for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanChaltiel/crosstable}
  \item Report bugs at \url{https://github.com/DanChaltiel/crosstable/issues}
}

}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
