% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{crosstable_peek_options}
\alias{crosstable_peek_options}
\title{See which \code{crosstable} option is currently set.}
\usage{
crosstable_peek_options(keep_null = FALSE)
}
\arguments{
\item{keep_null}{set to TRUE to get a list}
}
\value{
A named list of crosstable options
}
\description{
See which \code{crosstable} option is currently set.
}
