% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{crosstable_reset_options}
\alias{crosstable_reset_options}
\title{Reset all \code{crosstable} options.}
\usage{
crosstable_reset_options(quiet = FALSE)
}
\arguments{
\item{quiet}{set to \code{TRUE} to remove the message.}
}
\value{
Nothing, called for its side effects
}
\description{
Reset all \code{crosstable} options.
}
