% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{ct_bind_cols}
\alias{ct_bind_cols}
\title{Combine crosstables}
\usage{
ct_bind_cols(...)
}
\arguments{
\item{...}{crosstable objects}
}
\value{
a wider crosstable
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Combine multiple crosstables into one. This makes it possible to have multi-column
crosstables with intra-group margins. A better solution will be found later.
}
\examples{
ct1 = mtcars2 \%>\%
crosstable(c(disp, cyl), by=c(am, vs),
           margin=c("row", "col"))

ct2 = mtcars2 \%>\%
  dplyr::mutate(vs="Total") \%>\%
  crosstable(c(disp, cyl), by=c(am, vs),
             margin=c("row", "col"))

ct_bind_cols(ct1, ct2) \%>\%
  af(header_show_n=TRUE)
}
