% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctftable.R
\name{ctftable}
\alias{ctftable}
\title{Create a contingency table with summary statistics using ftable}
\usage{
ctftable(
  ...,
  data = parent.frame(),
  row.vars = NULL,
  col.vars = NULL,
  stats = "count"
)
}
\description{
This function will create a contingency table with summary statistics
using bind.tales and ftable
}
\examples{
ctftable(Freq ~ Type + Cont, stats=c("count", "row"), data=MASS::housing)

}
