% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{diff.mean.boot}
\alias{diff.mean.boot}
\title{Effect measure for association between one continuous and one categorical variable}
\usage{
\method{diff}{mean.boot}(x, g, conf.level = 0.95, R = 500)
}
\arguments{
\item{x}{vector}

\item{conf.level}{confidence interval level}

\item{R}{number of bootstrap replication}

\item{y}{another vector}
}
\value{
a list with five componments
}
\description{
Effect measure for association between one continuous and one categorical variable
}
