% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{diff.mean.student}
\alias{diff.mean.student}
\title{Effect measure for association between one continuous and one categorical variable}
\usage{
\method{diff}{mean.student}(x, g, conf.level = 0.95)
}
\arguments{
\item{x}{vector}

\item{conf.level}{confidence interval level}

\item{y}{another vector}
}
\value{
a list with five componments
}
\description{
Effect measure for association between one continuous and one categorical variable
}
