% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{diff_mean_student}
\alias{diff_mean_student}
\title{Effect measure for association between one continuous and one categorical variable}
\usage{
diff_mean_student(x, g, conf.level = 0.95)
}
\arguments{
\item{x}{vector}

\item{g}{another vector}

\item{conf.level}{confidence interval level}
}
\value{
a list with five componments
}
\description{
Effect measure for association between one continuous and one categorical variable
}
