% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{display_test}
\alias{display_test}
\title{Default function to display a test result}
\usage{
display_test(test, digits = 4, method = TRUE)
}
\arguments{
\item{test}{test}

\item{digits}{number of digits}

\item{method}{display method}
}
\value{
a string
}
\description{
Default function to display a test result
}
\author{
Dan Chaltiel
}
