% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_relative_risk}
\alias{effect_relative_risk}
\title{Effect measure for association between two factors}
\usage{
effect_relative_risk(x, y, conf.level = 0.95)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}

\item{conf.level}{confidence interval level}
}
\value{
a list with five componments
}
\description{
Effect measure for association between two factors
}
\seealso{
#https://stats.stackexchange.com/a/336624/81974
}
