% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_survival}
\alias{effect_survival}
\alias{effect_survival_coxph}
\title{Effect measure for association between one consored variable and one categorical variable}
\usage{
effect_survival_coxph(formula, conf_level = 0.95)
}
\arguments{
\item{formula}{a formula}

\item{conf_level}{the confidence level required}
}
\value{
a list with two componments: p.value and method
}
\description{
Effect measure for association between one consored variable and one categorical variable
}
\author{
David Hajage
}
