% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.r
\name{elements}
\alias{elements}
\title{Return elements in cbind, or x}
\usage{
elements(x)
}
\arguments{
\item{x}{}
}
\description{
Return elements in cbind, or x
}
\author{
David Hajage
}
\keyword{internal}

