% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.matrix.r
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(x, nrow, ncol, what = NULL, drop = TRUE)
}
\arguments{
\item{x}{x}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{what}{what}

\item{drop}{drop}
}
\description{
expand
}
\keyword{internal}
