% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{FlexCrossTable-ReporteRs}
\alias{FlexCrossTable-ReporteRs}
\alias{FlexCrossTable}
\title{Create a FlexTable object from a table made by the cross function}
\usage{
FlexCrossTable(crosstable, compact = FALSE, id = ".id",
  variable = "variable", value = "value", p = "p")
}
\arguments{
\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}

\item{id}{name of the 'id' column}

\item{variable}{name of the 'variable' column}

\item{value}{name of the 'value' column}

\item{p}{name of the 'p' column}
}
\value{
A \code{FlexTable} object (see \code{ReporteRs} package)
}
\description{
Create a FlexTable object from a table made by the cross function
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
FlexCrossTable(mytable)
FlexCrossTable(mytable, TRUE)
}
}
\author{
David Hajage
}
\keyword{univar}
