% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{FlexCrossTable}
\alias{FlexCrossTable}
\title{Create a FlexTable object from a table made by the cross function}
\usage{
FlexCrossTable(crosstable, compact = FALSE)
}
\arguments{
\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}
}
\value{
A \code{FlexTable} object
}
\description{
Create a FlexTable object from a table made by the cross function
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
FlexCrossTable(mytable)
FlexCrossTable(mytable, TRUE)
}
}
\author{
David Hajage
}
\keyword{univar}

