% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{generate_autofit_macro}
\alias{generate_autofit_macro}
\title{Generate a macro file for autofitting}
\usage{
generate_autofit_macro()
}
\value{
nothing
}
\description{
This function generates a file that can be imported into MS Word in order to use a macro for autofitting all tables in a document at once. This macro file should be imported only once per computer.
}
\section{Installation}{

\itemize{
\item Run \code{generate_autofit_macro()} in \code{R} to generate the file \code{crosstable_autofit.bas} in your working directory.
\item In MS Word, press Alt+F11 to open the VB Editor.
\item In the Editor, go to \code{File} > \code{Import} or press \code{Ctrl+M} to open the import dialog, and import \code{crosstable_autofit.bas}. There should now be a "CrosstableMacros" module in the "Normal" project.
\item Run the macro, either from the VB Editor or from \code{View} > \code{Macros} > \verb{View Macros} > \code{Run}.
}
}

