% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_label}
\alias{get_label}
\title{Get variable labels, or return a default value}
\usage{
get_label(
  x,
  default = names(x),
  object = FALSE,
  recursive = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{x}{A labelled object.
If \code{x} is a list or data frame, labels can be retrieved recursively from
its elements.}

\item{default}{Value returned when no label is found.
Defaults to \code{names(x)}.}

\item{object}{Logical.
If \code{TRUE}, and \code{x} is a list or data frame, return the label of \code{x} itself
instead of the labels of its children.}

\item{recursive}{Logical.
If \code{TRUE}, recurse into nested lists. Otherwise, labels are only retrieved
from the first level.}

\item{simplify}{Logical.
If \code{TRUE} and \code{object = FALSE}, simplify the result to a character vector
when possible. Otherwise, return a list.}
}
\value{
A character vector when \code{simplify = TRUE}, or a list otherwise.
}
\description{
Returns the \code{"label"} attribute of \code{x} when available.
If no label is found, \code{default} is returned instead.
}
\details{
If \code{x} is a list or a data frame, \code{get_label()} returns the labels of its
elements by default. Use \code{object = TRUE} to retrieve the label of the object
itself instead.
}
\examples{
xx=mtcars2 \%>\%
  set_label("The mtcars2 dataset", object=TRUE)
xx$cyl=remove_label(xx$cyl)

#vectors
get_label(xx$mpg) #label="Miles/(US) gallon"
get_label(xx$cyl) #default to NULL (since names(xx$cyl)==NULL)
get_label(xx$cyl, default="Default value")

#data.frames
get_label(xx)
get_label(xx, object=TRUE)
data.frame(name=names(xx), label=get_label(xx, default=NA)) #cyl is NA

#lists
get_label(list(xx$cyl, xx$mpg))          #cyl is NA
get_label(list(foo=xx$cyl, bar=xx$mpg))  #default to names
get_label(list(foo=xx$cyl, bar=xx$mpg), default="Default value")
}
\seealso{
\code{\link[=set_label]{set_label()}}, \code{\link[=import_labels]{import_labels()}}, \code{\link[=remove_label]{remove_label()}}, \code{\link[Hmisc:label]{Hmisc::label()}}, \code{\link[expss:var_lab]{expss::var_lab()}}
}
\author{
Dan Chaltiel
}
