% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{import_labels}
\alias{import_labels}
\title{Import labels from a dataset}
\usage{
import_labels(
  .tbl,
  data_label,
  name_from = "name",
  label_from = "label",
  verbose = TRUE
)
}
\arguments{
\item{.tbl}{the data.frame to labellize}

\item{data_label}{a data.frame from which to import labels}

\item{name_from}{in \code{data_label}, which column to get the variable name}

\item{label_from}{in \code{data_label}, which column to get the variable label}

\item{verbose}{if TRUE, displays a warning if a variable name is not found in \code{data_label}}
}
\description{
Import labels from a dataset
}
\examples{
iris_label = tibble::tibble(name=c("Sepal.Length", "Sepal.Width",
                                   "Petal.Length", "Petal.Width", "Species"),
                            label=c("Length of Sepals", "Width of Sepals",
                                    "Length of Petals", "Width of Petals", "Specie name"))
iris \%>\% 
  import_labels(iris_label) \%>\% 
  crosstable
}
