% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{is.crosstable}
\alias{is.crosstable}
\title{Test if an object is a crosstable}
\usage{
is.crosstable(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the \code{crosstable} class.
}
\description{
Test if an object is a crosstable
}
