% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remix.r
\name{is.numeric.and.not.surv}
\alias{is.numeric.and.not.surv}
\title{test}
\usage{
\method{is.numeric}{and.not.surv}(x)
}
\arguments{
\item{x}{x}
}
\description{
test
}
\author{
David Hajage
}
\keyword{internal}

