% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mediqr}
\alias{mediqr}
\title{Return median and IQR}
\usage{
mediqr(x, na.rm = TRUE, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{\code{TRUE} as default}

\item{dig}{number of digits}
}
\description{
Return median and IQR
}
\keyword{internal}

