% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.r
\name{Min}
\alias{Min}
\title{Return the min (and have formals)}
\usage{
Min(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{Remove NA?}
}
\description{
Return the min (and have formals)
}
\author{
David Hajage
}
\keyword{univar}
