% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moystd}
\alias{moystd}
\title{Return mean and sd}
\usage{
moystd(x, na.rm = TRUE, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{\code{TRUE} as default}

\item{dig}{number of digits}
}
\description{
Return mean and sd
}
\keyword{internal}

