% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{N}
\alias{N}
\title{Return the number of non NA observations}
\usage{
N(x)
}
\arguments{
\item{x}{a vector}
}
\value{
integer, number of non NA observations
}
\description{
Return the number of non NA observations
}
\author{
David Hajage
}
