% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{or.row.by.col}
\alias{or.row.by.col}
\title{Effect measure for association between two factors}
\usage{
or.row.by.col(x, y, conf.level = 0.95)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}

\item{conf.level}{confidence interval level}
}
\value{
a list with five componments
}
\description{
Effect measure for association between two factors
}
