% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.r
\name{parse_data}
\alias{parse_data}
\title{Parse data}
\usage{
parse_data(formula, data)
}
\arguments{
\item{formula}{formula (character)}

\item{data}{data}
}
\description{
Parse data
}
\author{
David Hajage
}
\keyword{internal}

