% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula}
\usage{
parse_formula(formula, varnames)
}
\arguments{
\item{formula}{formula (character)}

\item{varnames}{variables names}
}
\description{
Parse a formula
}
\author{
David Hajage
}
\keyword{internal}

