% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{peek}
\alias{peek}
\title{Open a \code{crosstable} in a temporary Word document}
\usage{
peek(x, ...)
}
\arguments{
\item{x}{a crosstable}

\item{...}{passed on to \code{as_flextable.crosstable()}}
}
\description{
This eases copy-pasting
}
