% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_crosstable}
\alias{pivot_crosstable}
\title{Pivot a crosstable}
\usage{
pivot_crosstable(ct)
}
\arguments{
\item{ct}{a crosstable}
}
\description{
Pivot a crosstable so the \code{variable} column is spread across its values.
}
\examples{
ct = crosstable(mtcars2, c(mpg, drat, wt, qsec), by=am)
t_ct = t(ct)
as_flextable(t_ct)
}
