% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.r
\name{plim}
\alias{plim}
\title{format p values}
\usage{
plim(p, digits = 4)
}
\arguments{
\item{p}{p values}

\item{digits}{number of digits}
}
\value{
formated p values
}
\description{
format p values
}
\author{
David Hajage
}

