% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.cross}
\alias{print.cross}
\title{Print a cross object}
\usage{
\method{print}{cross}(x, ...)
}
\arguments{
\item{x}{a cross object}

\item{...}{other arguments passed to nothing
which has no effect)}
}
\description{
Print cross object using ascii package
}
\author{
David Hajage
}
\keyword{univar}

