% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind.list}
\alias{rbind.list}
\title{Rbind for list}
\usage{
\method{rbind}{list}(l, preserve.attr = FALSE)
}
\arguments{
\item{l}{a list}
}
\description{
Rbind for list
}
\author{
David Hajage
}
\keyword{internal}
