% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind_crosstable}
\alias{rbind_crosstable}
\title{Rbind for crosstables}
\usage{
rbind_crosstable(...)
}
\arguments{
\item{...}{Some crosstable objects}
}
\value{
A \code{docx} object
}
\description{
Rbind for crosstables
}
\author{
David Hajage
}
