% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix.r
\name{regroup}
\alias{regroup}
\title{Regroup factors with factors, and numerical variables with numerical variables}
\usage{
regroup(vars, numdata, catdata, survdata)
}
\arguments{
\item{vars}{vars}

\item{numdata}{numdata}

\item{catdata}{catdata}

\item{survdata}{survdata}
}
\description{
Regroup factors with factors, and numerical variables with numerical variables
}
\author{
David Hajage
}
\keyword{internal}
