% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.r
\name{remove_blank}
\alias{remove_blank}
\title{Remove blank in a formula}
\usage{
remove_blank(formula)
}
\arguments{
\item{formula}{formula (character)}
}
\description{
Remove blank in a formula
}
\author{
David Hajage
}
\keyword{internal}

