% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{remove_label}
\alias{remove_label}
\title{Remove all label attributes.}
\usage{
remove_label(x)
}
\arguments{
\item{x}{object to unlabel}
}
\description{
Remove all label attributes.
}
\examples{
crosstable(remove_label(mtcars2))
crosstable(mtcars2)
}
\seealso{
\link{get_label}, \link{set_label}, \link{import_labels}, \link[expss:unlab]{expss::unlab}
}
