% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_dataframe_with_labels}
\alias{rename_dataframe_with_labels}
\title{Changes}
\usage{
rename_dataframe_with_labels(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Changes
}
\examples{
library(dplyr)
mtcars2 \%>\% 
  select(1:5) \%>\% 
  rename_dataframe_with_labels()
}
