% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{rr.col.by.row}
\alias{rr.col.by.row}
\title{Effect measure for association between two factors}
\usage{
rr.col.by.row(x, y, conf.level = 0.95)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}

\item{conf.level}{confidence interval level}
}
\value{
a list with five componments
}
\description{
Effect measure for association between two factors
}
