% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.r
\name{simple.table}
\alias{simple.table}
\title{Create a FlexTable object with a very simple style}
\usage{
simple.table(dataset, add.rownames = FALSE)
}
\arguments{
\item{dataset}{a data.frame}

\item{add.rownames}{a logical}
}
\value{
a FlexTable object
}
\description{
Create a FlexTable object with a very simple style
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- simple.table(head(iris))
}
}
\author{
David Hajage
}
\keyword{univar}
