% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_wrap2}
\alias{str_wrap2}
\title{Small improvement around stringr::str_wrap in case there is no whitespace}
\usage{
str_wrap2(x, width = getOption("crosstable_wrap", 70), ...)
}
\arguments{
\item{x}{character vector of strings to reformat}

\item{width}{target line width}

\item{...}{passed on to stringr::str_wrap}
}
\description{
Small improvement around stringr::str_wrap in case there is no whitespace
}
\examples{
set.seed(0)
x=sample(iris$Species, 10)
x \%>\% paste(collapse="") \%>\% str_wrap2(20) \%>\% cat
x \%>\% paste(collapse=" ") \%>\% str_wrap2(20) \%>\% cat
}
\keyword{internal}
