% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.by.r
\name{summarize.by}
\alias{summarize.by}
\title{Compute summary statistics according to a factor}
\usage{
summarize.by(x, by, funs = c(mean, sd, quantile, n, na), ...,
  showNA = c("no", "ifany", "always"), total = FALSE, digits = 2,
  test = FALSE, test.summarize = test.summarize.auto,
  show.test = display.test, plim = 4, show.method = TRUE)
}
\arguments{
\item{x}{numeric object}

\item{by}{factor}

\item{funs}{functions}

\item{...}{passed to funs}

\item{showNA}{showNA}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.summarize}{test.summarize}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}
}
\description{
Compute summary statistics according to a factor
}
\author{
David Hajage
}
\keyword{internal}
