% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize.data.frame.by}
\alias{summarize.data.frame.by}
\title{Compute summary statistics according to a factor (data.frame input)}
\usage{
summarize.data.frame.by(df, by, funs = c(mean, sd, quantile, n, na), ...,
  showNA = c("no", "ifany", "always"), total = FALSE, digits = 2,
  test = FALSE, test.summarize = test.summarize.auto,
  show.test = display.test, plim = 4, show.method = TRUE, label = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{by}{data.frame}

\item{funs}{fuctions}

\item{...}{passed to funs}

\item{showNA}{showNA}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.summarize}{test.summarize}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{label}{label}
}
\description{
Compute summary statistics according to a factor (data.frame input)
}
\author{
David Hajage
}
\keyword{internal}

