% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarize.data.frame}
\alias{summarize.data.frame}
\title{Compute summary statistics (data.frame input)}
\usage{
summarize.data.frame(df, funs = c(mean, sd, quantile, n, na), ...,
  digits = 2, label = FALSE)
}
\arguments{
\item{df}{a data.frame}

\item{funs}{functions}

\item{...}{passed to funs}

\item{digits}{number of digits}

\item{label}{add labels}
}
\description{
Compute summary statistics (data.frame input)
}
\author{
David Hajage
}
\keyword{internal}

