% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{Compute summary statistics}
\usage{
summarize(x, funs = c(mean, sd, quantile, n, na), ..., digits = 2)
}
\arguments{
\item{x}{numeric object}

\item{funs}{functions}

\item{...}{passed to funs}

\item{digits}{number of digits}
}
\description{
Compute summary statistics
}
\author{
David Hajage
}
\keyword{internal}

