% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.r
\name{survival}
\alias{survival}
\title{Compute survival}
\usage{
survival(surv, by = NULL, times = NULL, followup = FALSE,
  total = FALSE, digits = 2, test = FALSE,
  test.survival = test.survival.logrank, show.test = display.test,
  plim = 4, show.method = TRUE, effect = FALSE,
  effect.survival = effect.survival.coxph,
  show.effect = display.effect, conf.level = 0.95, label = FALSE)
}
\arguments{
\item{surv}{a Surv object}

\item{by}{by}

\item{times}{times}

\item{followup}{followup}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.survival}{test.survival}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{effect}{effect}

\item{effect.survival}{effect.survival}

\item{show.effect}{show.effect}

\item{conf.level}{conf.level}

\item{label}{label}
}
\description{
Compute survival
}
\author{
David Hajage
}
\keyword{internal}
