% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tabular.data.frame}
\alias{tabular.data.frame}
\title{Compute a contingency table (data.frame input)}
\usage{
tabular.data.frame(dfx, dfy, margin = 0:2, showNA = c("no", "ifany",
  "always"), total = FALSE, digits = 2, test = FALSE,
  test.tabular = test.tabular.auto, show.test = display.test, plim = 4,
  show.method = TRUE, label = FALSE)
}
\arguments{
\item{dfx}{data.frame}

\item{dfy}{data.frame}

\item{margin}{margin}

\item{showNA}{showNA}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.tabular}{test.tabular}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{label}{label}
}
\description{
Compute a contingency table (data.frame input)
}
\author{
David Hajage
}
\keyword{internal}

