% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.r
\name{tabular}
\alias{tabular}
\title{Compute a contingency table}
\usage{
tabular(x, y, showNA = c("no", "ifany", "always"), margin = 0:2,
  total = FALSE, digits = 2, test = FALSE,
  test.tabular = test.tabular.auto, show.test = display.test,
  plim = 4, show.method = TRUE, effect = FALSE,
  effect.tabular = or.row.by.col, conf.level = 0.95,
  show.effect = display.effect)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{showNA}{showNA}

\item{margin}{margin}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.tabular}{test.tabular}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{effect}{effect}

\item{effect.tabular}{effect.tabular}

\item{conf.level}{conf.level}

\item{show.effect}{show.effect}
}
\description{
Compute a contingency table
}
\author{
David Hajage
}
\keyword{internal}
