% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.summarize.contrasts.lin}
\alias{test.summarize.contrasts.lin}
\title{Test for linear trend across ordered factor with contrasts}
\usage{
test.summarize.contrasts.lin(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{ordered factor}
}
\value{
a list with two componments: p.value and method
}
\description{
Test for linear trend across ordered factor with contrasts
}
\examples{
library(dplyr)
my_test_args=crosstable_test_args()
my_test_args$test.summarize = test.summarize.contrasts.lin
iris \%>\%
  mutate(Petal.Width.qt = paste0("Q", ntile(Petal.Width, 5)) \%>\% ordered()) \%>\%
  crosstable(Petal.Length ~ Petal.Width.qt, test=TRUE, test_args = my_test_args)
}
\author{
Dan Chaltiel
}
