% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.r
\name{test.summarize.kruskal}
\alias{test.summarize.kruskal}
\title{test for mean comparison}
\usage{
test.summarize.kruskal(x, g)
}
\arguments{
\item{x}{vector}

\item{g}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a kruskal.test.
}
\author{
David Hajage
}

