% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.summarize.oneway.equalvar}
\alias{test.summarize.oneway.equalvar}
\title{test for mean comparison}
\usage{
test.summarize.oneway.equalvar(x, g)
}
\arguments{
\item{x}{vector}

\item{g}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a oneway.test. with var.equal = TRUE
}
\author{
David Hajage
}
