% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.r
\name{test.summarize.oneway.unequalvar}
\alias{test.summarize.oneway.unequalvar}
\title{test for mean comparison}
\usage{
test.summarize.oneway.unequalvar(x, g)
}
\arguments{
\item{x}{vector}

\item{g}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a oneway.test. with var.equal = FALSE
}
\author{
David Hajage
}

