% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.survival.logrank}
\alias{test.survival.logrank}
\title{test for survival comparison}
\usage{
test.survival.logrank(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a logrank test
}
\author{
David Hajage
}

