% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.tabular.auto}
\alias{test.tabular.auto}
\title{test for contingency table}
\usage{
test.tabular.auto(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a chisq.test, a chisq.test with correction of continuity
or a fisher test as appropriate
}
\author{
David Hajage
}

