% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_correlation_auto}
\alias{test_correlation_auto}
\title{test for correlation coefficients}
\usage{
test_correlation_auto(x, by, method)
}
\arguments{
\item{x}{vector}

\item{by}{another vector}

\item{method}{"pearson", "kendall", or "spearman"}
}
\value{
the correlation test with appropriate method
}
\description{
test for correlation coefficients
}
\author{
Dan Chaltiel, David Hajage
}
