% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_summarize_auto}
\alias{test_summarize_auto}
\title{test for mean comparison}
\usage{
test_summarize_auto(x, g)
}
\arguments{
\item{x}{vector}

\item{g}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a oneway.test (with equal or unequal variance) or a
kruskal.test as appropriate.
}
\author{
David Hajage, Dan Chaltiel
}
