% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_summarize_linear_contrasts}
\alias{test_summarize_linear_contrasts}
\title{Test for linear trend across ordered factor with contrasts}
\usage{
test_summarize_linear_contrasts(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{ordered factor}
}
\value{
a list with two components: p.value and method
}
\description{
Test for linear trend across ordered factor with contrasts
}
\examples{
library(dplyr)
my_test_args=crosstable_test_args()
my_test_args$test_summarize = test_summarize_linear_contrasts
iris \%>\%
  mutate(Petal.Width.qt = paste0("Q", ntile(Petal.Width, 5)) \%>\% ordered()) \%>\%
  crosstable(Petal.Length ~ Petal.Width.qt, test=TRUE, test_args = my_test_args)
}
\author{
Dan Chaltiel
}
