% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_survival_logrank}
\alias{test_survival_logrank}
\title{test for survival comparison}
\usage{
test_survival_logrank(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
a list with two components: p.value and method
}
\description{
Compute a logrank test
}
\author{
Dan Chaltiel, David Hajage
}
