% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_tabular_auto}
\alias{test_tabular_auto}
\title{test for contingency table}
\usage{
test_tabular_auto(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a chisq.test, a chisq.test with correction of continuity
or a fisher test as appropriate
}
\author{
David Hajage
}
