% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{transpose_crosstable}
\alias{transpose_crosstable}
\alias{t.crosstable}
\title{Transpose a crosstable}
\usage{
transpose_crosstable(x)

\method{t}{crosstable}(x)
}
\arguments{
\item{x}{a crosstable}
}
\description{
Pivot a crosstable so the \code{label} column is swapped with the \code{by} row.
This requires the \code{variable} column to be the same for every data column, like when all columns are numeric of when all columns are factors with the same levels
}
\examples{
ct = crosstable(mtcars2, c(mpg, drat, wt, qsec), by=am)
t_ct = t(ct)
as_flextable(t_ct)
}
