% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.r
\name{trim}
\alias{trim}
\title{Remove blancks at the begining and the end}
\usage{
trim(x)
}
\arguments{
\item{x}{x}
}
\description{
Remove blancks at the begining and the end
}
\author{
David Hajage
}
\keyword{internal}
