% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{write_and_open}
\alias{write_and_open}
\title{Alternative to default \code{officer} print() function. Write the file and try to open it right away.}
\usage{
write_and_open(doc, docx.file)
}
\arguments{
\item{doc}{the docx object}

\item{docx.file}{the name of the target file}
}
\description{
As it tests if the file is writable, this function also prevents \code{\link[officer:::print.rdocx]{officer::::print.rdocx()}} to abort the RStudio session.
}
\examples{
\dontrun{
library(officer)
library(crosstable)
mytable = crosstable(mtcars2)
doc = read_docx() \%>\% 
    body_add_crosstable(mytable)
write_and_open(doc, "example.docx")
}
}
\author{
Dan Chaltiel
}
