% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_crosstable.r
\name{addCrossTable}
\alias{addCrossTable}
\title{Deprecated
add a table made by the cross function into a ReporteRs document}
\usage{
addCrossTable(doc, crosstable, compact = FALSE, id = ".id",
  variable = "variable", value = "value", p = "p")
}
\arguments{
\item{doc}{a \code{docx} object created by \code{docx} function (see \code{ReporteRs} package)}

\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}

\item{id}{name of the 'id' column}

\item{variable}{name of the 'variable' column}

\item{value}{name of the 'value' column}

\item{p}{name of the 'p' column}
}
\value{
A \code{docx} object
}
\description{
**Deprecation** : Since \code{ReporteRs} is deprecated and not available on CRAN, please use \code{officer} and \code{biostats2::body_add_crosstable} instead.
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
doc <- docx()
doc <- addCrossTable(doc, mytable)
doc <- addPageBreak(doc)
doc <- addCrossTable(doc, mytable, TRUE)
dfile <- "test_doc.docx"
print(doc, target = dfile)
shell.exec(dfile)
}
}
\author{
David Hajage
}
\keyword{univar}
