% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{apply_labels}
\alias{apply_labels}
\title{Batch set variable labels}
\usage{
apply_labels(data, ..., fn, warn_missing = FALSE)
}
\arguments{
\item{data}{data.frame/list}

\item{...}{named arguments, as \code{colname="label"}}

\item{fn}{alternatively, a function to be applied to all existing labels.}

\item{warn_missing}{if TRUE, throw a warning if some names are missing}
}
\value{
An object of the same type as \code{data}, with labels
}
\description{
This function is a copycat of from expss package v0.10.7 (slightly modified) to avoid having to depend on expss. See \code{\link[expss:apply_labels]{expss::apply_labels()}} for more documentation. Note that this version is not compatible with \code{data.table}.
}
\examples{
iris \%>\%
  apply_labels(Sepal.Length="Length of Sepal",
               Sepal.Width="Width of Sepal") \%>\%
  crosstable()
iris2 \%>\%
  apply_labels(fn=tolower) \%>\%
  crosstable()
}
\author{
Dan Chaltiel
}
