% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\docType{import}
\name{as_flextable.crosstable}
\alias{as_flextable.crosstable}
\alias{ctf}
\alias{cross_to_flextable}
\alias{as_flextable}
\title{Turns a \code{crosstable} object into a \code{flextable}}
\usage{
\method{as_flextable}{crosstable}(
  x,
  autofit = TRUE,
  compact = FALSE,
  show_test_name = TRUE,
  by_header = NULL,
  keep_id = FALSE,
  generic_labels = list(id = ".id", variable = "variable", value = "value", total =
    "Total", label = "label", test = "test", effect = "effect"),
  ...
)
}
\arguments{
\item{x}{the result of \code{\link[=crosstable]{crosstable()}}}

\item{autofit}{whether to use \code{flextable::autofit} on the table}

\item{compact}{whether to compact the table}

\item{show_test_name}{in the \code{test} column, show the test name}

\item{by_header}{a string to override the \code{by} header}

\item{keep_id}{whether to keep the \code{.id} column}

\item{generic_labels}{names of the crosstable default columns}

\item{...}{unused}
}
\section{Methods (by class)}{
\itemize{
\item \code{crosstable}: Turns a \code{crosstable} object into a formatted \code{flextable}.\cr You can also use the alias \code{cross_to_flextable} or its shortcut \code{ctf()} which has the same arguments.
}}

\examples{
#Crosstables
library(crosstable)
library(dplyr)
crosstable(mtcars2) \%>\% as_flextable()
crosstable(mtcars2, by=vs, test = TRUE) \%>\% as_flextable()
crosstable(esoph, by=tobgp, test = TRUE) \%>\% as_flextable()
crosstable(iris, by=Species, test = TRUE) \%>\% as_flextable(autofit=FALSE)
crosstable(iris, by=Species, test = TRUE) \%>\% as_flextable(compact=TRUE)
crosstable(iris) \%>\% as_flextable(compact=TRUE, autofit=TRUE)

#Renaming
crosstable(iris, by=Species, total="both", test=TRUE, effect=TRUE) \%>\%
   rename(ID=.id, math=variable, Tot=Total, lab=label, pval=test, fx=effect) \%>\%
   as_flextable(by_header = "The specie", 
                      generic_labels=list(id = "ID", variable = "math", total="Tot", 
                                          label = "lab", test = "pval", effect="fx"))
}
\seealso{
\code{\link[=crosstable]{crosstable()}}, \code{\link[flextable:flextable]{flextable::flextable()}}
}
\author{
Dan Chaltiel
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flextable}{\code{\link[flextable]{as_flextable}}}
}}

