% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_crosstable}
\alias{body_add_crosstable}
\title{body_add_crosstable}
\usage{
body_add_crosstable(doc, x, ...)
}
\arguments{
\item{doc}{a \code{rdocx} object created by \code{read_docx} function (see \code{officer} package)}

\item{x}{a \code{crosstable}}

\item{...}{further arguments for \code{as_flextable.crosstable}}
}
\description{
\code{\link[=body_add_crosstable]{body_add_crosstable()}} adds such a \code{flextable} an \code{officer} document.
}
\examples{
#Officer
library(officer)
library(dplyr)
mytable = crosstable(mtcars2)
doc = read_docx() \%>\% 
    body_add_crosstable(mytable) \%>\% 
    body_add_break \%>\% 
    body_add_crosstable(mytable, compact=TRUE)

\dontrun{
dfile = "examples\\\\example_doc.docx"
print(doc, target = dfile)
shell.exec(dfile)
}
}
