% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{body_add_crosstable}
\alias{body_add_crosstable}
\title{add a table made by the cross function into an officer document}
\usage{
body_add_crosstable(doc, crosstable, compact = FALSE, auto.fit = FALSE,
  id = ".id", variable = "variable", label = "label",
  value = "value", p = "p", show.test.name = F)
}
\arguments{
\item{doc}{a \code{rdocx} object created by \code{read_docx} function (see \code{officer} package)}

\item{crosstable}{the result of \code{cross} function}

\item{compact}{whether to compact the table}

\item{auto.fit}{whether to \code{flextable::autofit} the table}

\item{id}{name of the 'id' column}

\item{variable}{name of the 'variable' column}

\item{label}{name of the 'label' column}

\item{value}{name of the 'value' column}

\item{p}{name of the 'p' column}

\item{show.test.name}{in the p column, show the test name}
}
\value{
A \code{rdocx} object
}
\description{
add a table made by the cross function into an officer document
}
\examples{
library(dplyr) #for the pipe operator
library(officer)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
mytable <- cross(cbind(...) ~ Species, iris, test = TRUE)
mytable <- cross(cbind(...) ~ ., esoph)
doc <- read_docx() \%>\% 
    body_add_crosstable(mytable) \%>\% 
    body_add_break \%>\% 
    body_add_crosstable(mytable, TRUE) \%>\% 
    body_add_break \%>\% 
    body_add_crosstable(compact(mytable))

\dontrun{
dfile <- "test_doc.docx"
print(doc, target = dfile)
shell.exec(dfile)
}
}
\author{
Dan Chaltiel
}
