% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_crosstable_footnote}
\alias{body_add_crosstable_footnote}
\title{Adds a standard footnote explaining the abbreviations used in a crosstable}
\usage{
body_add_crosstable_footnote(doc)
}
\arguments{
\item{doc}{a \code{rdocx} object}
}
\value{
The docx object \code{doc}
}
\description{
Use it below \code{\link[=body_add_crosstable]{body_add_crosstable()}}.
Footnote: Med: median, IQR: interquartile range, Std: standard deviation. Percentages are expressed in column.
}
\author{
Dan Chaltiel
}
