% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_normal}
\alias{body_add_normal}
\alias{body_add_glued}
\title{Add a new paragraph with a Normal style, inserting variables with \code{base::paste}}
\usage{
body_add_normal(doc, ...)

body_add_glued(doc, x, ...)
}
\arguments{
\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}

\item{...}{one or several character strings, collapsed into a paragraph with \code{base::paste}}

\item{x}{the string with \code{glue::glue} patterns (Expressions enclosed by braces will be evaluated as R code)}

\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}

\item{...}{parameters to be passed to \code{glue::glue}}
}
\value{
a new doc object

a new doc object
}
\description{
Add a new paragraph with a Normal style, inserting variables with \code{base::paste}

Add a new paragraph with a Normal style, inserting variables with \code{glue::glue}
}
\examples{
\dontrun{
library(officer)
library(biostat2)
library(dplyr)
doc = read_docx()
doc = doc \%>\% body_add_normal("La table iris a ", ncol(iris), " colonnes.")
}
\dontrun{
library(officer)
library(biostat2)
library(dplyr)
doc = read_docx()
doc = body_add_glued(doc, "La table iris a {ncol(iris)} colonnes.")
}
}
\author{
Dan Chaltiel

Dan Chaltiel
}
