% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_normal}
\alias{body_add_normal}
\title{Add a new paragraph with a Normal style}
\usage{
body_add_normal(doc, ...)
}
\arguments{
\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}

\item{...}{one or several character strings to collapse into a paragraph}
}
\value{
a new doc object
}
\description{
Add a new paragraph with a Normal style
}
\examples{
\dontrun{
library(biostat2)
library(ReporteRs)
doc <- body_add_normal(doc, "La table iris a ", ncol(iris), " colonnes.")
}
}
\author{
Dan Chaltiel
}
\keyword{univar}
