% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_legend.r
\name{body_add_table_legend}
\alias{body_add_table_legend}
\title{Add a table legend}
\usage{
body_add_table_legend(x, legend, legend_style = "table title",
  style = "strong", seqfield = "SEQ Table \\\\* Arabic")
}
\arguments{
\item{x}{a docx object}

\item{legend}{the table legend}

\item{legend_style}{may depend on the docx template}

\item{style}{the legend style (strnog, italic...)}

\item{seqfield}{to figure this out, in a docx file, insert a table legend, right click on the inserted number and select "Toggle Field Codes". This argument should be the value of the field, with extra escaping.}
}
\description{
Add a table legend
}
