% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_table_legend}
\alias{body_add_table_legend}
\alias{body_add_figure_legend}
\title{Add a table legend to an \code{officer} document}
\usage{
body_add_table_legend(
  x,
  legend,
  legend_style = "table title",
  style = "strong",
  seqfield = "SEQ Table \\\\* Arabic"
)

body_add_figure_legend(
  x,
  legend,
  legend_style = "graphic title",
  style = "strong",
  seqfield = "SEQ Figure \\\\* Arabic"
)
}
\arguments{
\item{x}{a docx object}

\item{legend}{the table legend}

\item{legend_style}{may depend on the docx template}

\item{style}{the legend style (strong, italic...)}

\item{seqfield}{to figure this out, in a docx file, insert a table legend, right click on the inserted number and select "Toggle Field Codes". This argument should be the value of the field, with extra escaping.}
}
\description{
Add a table legend to an \code{officer} document
}
\section{Warning}{

Sometimes, the legends added with \code{body_add_table_legend} have no numbers. In this case, you have to manualy update the references: select all (\kbd{Ctrl+\kbd{A}), then update (\kbd{F9}).
}

\examples{
library(officer)
library(dplyr) #for pipes
library(ggplot2)
p=quickplot(x=Sepal.Length, y=Sepal.Width, color=Species, data=iris)
x=read_docx() \%>\% 
    body_add_table_legend("Iris dataset") \%>\% 
    body_add_crosstable(crosstable(iris)) \%>\% 
    body_add_gg(p) \%>\% 
    body_add_figure_legend("Iris plot")
#print(x, "example.docx")
}
