% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_add_table_list}
\alias{body_add_table_list}
\alias{body_add_flextable_list}
\alias{body_add_crosstable_list}
\title{Add a list of tables}
\usage{
body_add_table_list(doc, l, fun = "title2", fun_after = NULL, ...)

body_add_flextable_list(...)

body_add_crosstable_list(...)
}
\arguments{
\item{doc}{a \code{rdocx} object, created by \code{\link[officer:read_docx]{officer::read_docx()}}}

\item{l}{a named list of tables (of class \code{crosstable}, \code{flextable}, or \code{data.frame}).}

\item{fun}{a function to be used before each table, most likely to add some kind of title. Should be of the form \verb{function(doc, .name)} where \code{.name} is the name of the current crosstable of the list. You can also pass \code{"title2"} to add the name as a title of level 2 between each table, \code{"newline"} to simply add a new line, or even NULL to not separate them (beware that the table might merge then).}

\item{fun_after}{same as \code{fun} but is used after each table.}

\item{...}{arguments passed on to \code{\link[=body_add_crosstable]{body_add_crosstable()}} or \code{\link[=body_add_flextable]{body_add_flextable()}}}
}
\value{
The docx object \code{doc}
}
\description{
Add a list of tables in an officer document. \code{crosstables} will be added using \code{\link[=body_add_crosstable]{body_add_crosstable()}} and \code{flextables} will be added using \code{\link[flextable:body_add_flextable]{flextable::body_add_flextable()}}. Plain dataframes will be converted to flextables.
}
\examples{
library(officer)
ctl = list(iris2=crosstable(iris2, 1),
           "Just a flextable"=flextable::flextable(mtcars2[1:5,1:5]),
           "Just a dataframe"=iris2[1:5,1:5])

myfun = function(doc, .name){
    doc \%>\%
        body_add_title(" This is table '{.name}' as a flex/crosstable", level=2) \%>\%
        body_add_normal("Here is the table:")
}
myfun2 = function(doc, .name){
  doc \%>\% body_add_table_legend(.name, bookmark=janitor::make_clean_names(.name))
}
read_docx() \%>\%
  body_add_title("Separated by subtitle", 1) \%>\%
  body_add_table_list(ctl, fun="title2") \%>\%
  body_add_title("Separated using a custom function", 1) \%>\%
  body_add_normal("You can therefore use bookmarks, for instance here are tables \\\\@ref(iris2), \\\\@ref#' (just_a_flextable) and \\\\@ref(just_a_dataframe).") \%>\%
  body_add_table_list(ctl, fun=myfun, fun_after=myfun2, body_fontsize=8) \%>\%
  write_and_open()
}
