% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_title}
\alias{body_add_title}
\title{Add a new title}
\usage{
body_add_title(x, value, level = 1, style = "heading")
}
\arguments{
\item{value}{a character string}

\item{level}{the level of the title. See \code{styles_info(doc)} to know the possibilities.}

\item{doc}{the doc object (created with the \code{read_docx} function of \code{officer} package)}
}
\value{
a new doc object
}
\description{
Add a new title
}
\examples{
\dontrun{
library(officer)
library(biostat2)
library(dplyr)
doc = read_docx()
doc = doc \%>\% 
   body_add_title(doc, "La table iris", 1) \%>\% 
   body_add_title(doc, "Description", 2) \%>\% 
   addNormal(doc, "La table iris a ", ncol(iris), " colonnes.")
}
}
\author{
Dan Chaltiel
}
