% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{body_replace_text_at_bkms}
\alias{body_replace_text_at_bkms}
\title{Replace text on several bookmarks at once}
\usage{
body_replace_text_at_bkms(doc, ..., envir = parent.frame())
}
\arguments{
\item{doc}{a \code{rdocx} object}

\item{...}{named}

\item{envir}{Environment to evaluate each expression in \code{glue()}.}
}
\value{
The docx object \code{doc}
}
\description{
Replace text on several bookmarks at once
}
\author{
Dan Chaltiel
}
