% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{compact}
\alias{compact}
\title{Compact the result of cross function}
\usage{
compact(x)
}
\arguments{
\item{x}{x}
}
\description{
Compact the result of cross function
}
\author{
David Hajage
}
\keyword{internal}

