% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix.r
\name{cross}
\alias{cross}
\title{Describe everything}
\usage{
cross(formula = cbind(...) ~ ., data = NULL, funs = c(` ` = mysummary),
  ..., margin = 0:2, total = FALSE, digits = 2, showNA = c("no",
  "ifany", "always"), method = c("pearson", "kendall", "spearman"),
  times = NULL, followup = FALSE, test = FALSE,
  test.summarize = test.summarize.auto,
  test.survival = test.survival.logrank, test.tabular = test.tabular.auto,
  show.test = display.test, plim = 4, show.method = TRUE,
  effect = FALSE, effect.summarize = diff.mean.auto,
  effect.tabular = or.row.by.col, effect.survival = effect.survival.coxph,
  conf.level = 0.95, label = FALSE, regroup = FALSE)
}
\arguments{
\item{formula}{a formula (see Details).}

\item{data}{a data.frame.}

\item{funs}{Functions used for describing numeric
variables. Vector (named or not): \code{c(fun1, fun2, fun3)} or
\code{c("fun1", "fun2", "fun3")}.}

\item{...}{further arguments, all passed to funs. For example
{na.rm = TRUE}}

\item{margin}{index, or vector of indices to indicate which
proportions should be computed in frequency tables (0: cell, 1:
row, 2: col).}

\item{total}{whether to add margins. Integers (\code{c(1, 2)}:
both row and col margins, 1: row margins, 2: col margins, 0: no
margins) or logical (\code{TRUE}: row and col margins,
\code{FALSE}: no margins).}

\item{digits}{number of digits}

\item{showNA}{whether to show NA (\code{c("no", "ifany", "always"}
like in \code{table()})}

\item{method}{a character string indicating which correlation
coefficient is to be used. One of \code{"pearson"},
\code{"kendall"}, or \code{"spearman"}, can be abbreviated.}

\item{times}{vector of times (see \code{?summary.survival} in
package \code{survival}).}

\item{followup}{whether to display follow-up time.}

\item{test}{whether to perform tests}

\item{test.summarize}{a function of two arguments (continuous
variable and grouping variable), used to compare continuous
variable. Returns a list of two components : \code{p.value} and
\code{method} (the test name). See \code{test.summarize.auto},
\code{test.summarize.kruskal},
\code{test.summarize.oneway.equalvar}, or
\code{test.summarize.unequalvar} for some examples of such
functions. Users can provide their own function.}

\item{test.survival}{a function of one argument (a formula), used
to compare survival estimations. Returns the same components as
created by \code{test.summarize}. See
\code{test.survival.logrank}. Users can provide their own
function.}

\item{test.tabular}{a function of two arguments (two categorical
variables), used to test association between two factors.  Returns
the same components as created by \code{test.summarize}. See
\code{test.tabular.auto} and \code{test.tabular.fisher}. Users can
provide their own function.}

\item{show.test}{function used to display the test result. See
\code{display.test}.}

\item{plim}{number of digits for the p value}

\item{show.method}{wether to display the test name (logical)}

\item{effect}{whether to compute a effect measure}

\item{effect.summarize}{a function of three arguments (continuous
variable, grouping variable and conf.level), used to compare continuous
variable. Returns a list of five components : \code{effect} (the effect value(s)),
\code{ci} (the matrix of confidence interval(s)), \code{effect.name} (the 
interpretiation(s) of the effect value(s)), \code{effect.type} (the description of 
the measure used) and \code{conf.level} (the confidence interval level). See 
\code{diff.mean.auto}, \code{diff.mean.student} or \code{diff.mean.boot} for some examples of such
functions. Users can provide their own function.}

\item{effect.tabular}{a function of three arguments (two categorical variables and conf.level) used to measure the associations between two factors.
Returns a list of five components : \code{effect} (the effect value(s)),
\code{ci} (the matrix of confidence interval(s)), \code{effect.name} (the 
interpretiation(s) of the effect value(s)), \code{effect.type} (the description of 
the measure used) and \code{conf.level} (the confidence interval level). See 
\code{or.row.by.col}, \code{rr.row.by.col}, \code{rd.row.by.col}, \code{or.col.by.row}, \code{rr.col.by.row}, or \code{rd.col.by.row} for some examples of such functions. Users can provide their own function.}

\item{effect.survival}{a function of two argument (a formula and conf.level), used
to measure the association between a consored and a factor. Returns the same components as
created by \code{effect.summarize}. See
\code{effect.survival.coxph}. Users can provide their own
function.}

\item{conf.level}{The desired confidence interval level}

\item{label}{whether to display labels of variables (using
\code{label} in package \code{Hmisc})}

\item{regroup}{whether to regroup numerics with numerics and
factors with factors in \code{cbind} (logical)}
}
\value{
A data.frame, or a list of data.frames.
}
\description{
A quick and easy function for describing datasets.
}
\note{
The formula has the following format: \code{x_1 + x_2 +
... ~ y_1 + y_2 + ...}

There are a couple of special variables: \code{...} represents all
other variables not used in the formula and \code{.}  represents
no variable, so you can do \code{formula = var1 ~ .}.

If \code{var1} is numeric, \code{var1 ~ .} produce a summary table
using \code{funs}. If \code{var1} is a factor, \code{var1 ~ .}
produce a frequency table. If \code{var1} is of class \code{Surv},
\code{var1 ~ .} produce a table with the estimates of survival at
\code{times}. If \code{var1} is numeric and \code{var2} is
numeric, \code{var1 ~ var2} produces a correlation correlation
coefficient. if \code{var1} is numeric and \code{var2} is a
factor, \code{var1 ~ var2} produce a summary table (using
functions in \code{funs}) according to the levels of
\code{var2}. If \code{var1} is a factor and \code{var2} is a
factor, \code{var1 ~ var2} produce a contingency table. If
\code{var1} is of class \code{Surv} and \code{var2} is a factor,
\code{var1 ~ var2} produce a table with the estimates of survival
for each level of \code{var2}.

You can group several variables together with \code{cbind(var1,
var2, var3)}: \code{var1}, \code{var2} and \code{var3} will be
grouped in the same table. \code{cbind(...)} works (ie regroups
all variables of the data.frame together). When a \code{cbind} is
in both sides of the formula, \code{cross} will do its best to
group everything in the same table, but only if it is possible...
}
\examples{

library(biostat2)
cross(data = iris)
cross(cbind(...) ~ ., iris[, sapply(iris, is.numeric)], funs = c(median, mad, min, max))
cross(cbind(Sepal.Length, I(Sepal.Width^2)) ~ Species, iris, funs = quantile, probs = c(1/3, 2/3))
cross(Sepal.Length + Sepal.Width ~ Petal.Length + Petal.Width, iris)
cross(cbind(Sepal.Length, Sepal.Width) ~ cbind(Petal.Length, Petal.Width), iris)
cross(... ~ ., esoph)
cross(alcgp ~ tobgp, esoph, test = TRUE)
library(survival)
cross(Surv(time, status) ~ x, data = aml)
}
\seealso{
\code{cast} (reshape) and \code{summary.formula} (Hmisc).
}
\author{
David Hajage, inspired by the design and the code of
  \code{summary.formula} (\code{Hmisc} package, FE Harrell) and
  \code{cast} (\code{reshape} package, H Wickham).
}
\keyword{univar}
