% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{cross_summary}
\alias{cross_summary}
\title{Summarize a numeric vector}
\usage{
cross_summary(x, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{dig}{number of digits}
}
\description{
Summarize a numeric vector with min, max, mean, sd, median, IQR, n and missings.
}
\examples{
cross_summary(iris$Sepal.Length)
}
