% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{Easily describe datasets}
\usage{
crosstable(
  data,
  .vars = NULL,
  ...,
  by = NULL,
  margin = c("row", "column", "cell", "all"),
  total = c("none", "row", "column", "both"),
  percent_digits = 2,
  showNA = c("ifany", "always", "no"),
  label = TRUE,
  funs = c(` ` = cross_summary),
  funs_arg = list(),
  test = FALSE,
  test_args = crosstable_test_args(),
  cor_method = c("pearson", "kendall", "spearman"),
  unique_numeric = 3,
  effect = FALSE,
  effect_args = crosstable_effect_args(),
  times = NULL,
  followup = FALSE
)
}
\arguments{
\item{data}{a data.frame}

\item{.vars}{the variables to describe. Can be a character or name vector, a tidyselect helper, a (lambda) function that returns a logical, or a formula. See examples or \code{vignette("crosstable-selection")} for more details.}

\item{...}{more variables to describe. Cannot be a lambda function nor a formula.}

\item{by}{the variable to group on. Character or name.}

\item{margin}{one of ["row", "column", "cell" or "all"] to indicate which proportions should be computed in frequency tables. Default to \code{row}.}

\item{total}{one of ["none", "row", "column" or "both"] to indicate whether to add margins. Default to \code{none}.}

\item{percent_digits}{number of digits for percentages}

\item{showNA}{whether to show NA in factors (one of \code{c("ifany", "always", "no")}, like in \code{table()})}

\item{label}{whether to show labels. Use either \code{\link[Hmisc:label]{Hmisc::label()}}, \code{\link[expss:var_lab]{expss::var_lab()}} or \code{expss::apply_labels} to add labels to the dataset columns.}

\item{funs}{functions to apply to numeric variables. Default to \code{\link{cross_summary}}.}

\item{funs_arg}{additionnal parameters for \code{funs}. With \code{cross_summary}, you can set \code{dig}, the number of decimal places.}

\item{test}{whether to perform tests}

\item{test_args}{See \code{\link{crosstable_test_args}} to override default testing behaviour.}

\item{cor_method}{one of ["pearson", "kendall", or "spearman"] to indicate which correlation coefficient is to be used.}

\item{unique_numeric}{the number of non-missing different levels a variable should have to be considered as numeric}

\item{effect}{whether to compute a effect measure}

\item{effect_args}{See \code{\link{crosstable_effect_args}} to override default behaviour.}

\item{times}{when using formula with \code{\link[survival:Surv]{survival::Surv()}} objects, which times to summarize}

\item{followup}{when using formula with \code{\link[survival:Surv]{survival::Surv()}} objects, whether to display follow-up time}
}
\description{
Easily describe datasets
}
\examples{
#whole table
crosstable(iris)
crosstable(mtcars)

#tidyselection, custom functions
library(tidyverse)
crosstable(mtcars2, ends_with("t"), starts_with("c"), by=vs, 
           funs=c(mean, quantile), funs_arg = list(probs=c(.25,.75)))

#margin and totals
crosstable(mtcars2, disp, vs, by=am, 
           margin=c("row", "col"), total = "both")

#predicate selection, correlation, testing
crosstable(mtcars2, is.numeric, by=hp, test=TRUE)

#lambda selection, effect calculation
crosstable(mtcars2, ~is.numeric(.x) && mean(.x)>50, by=vs, effect=TRUE)


#Survival data (using formula UI)
library(survival)
crosstable(aml, Surv(time, status) ~ x,times=c(0,15,30,150), followup=TRUE)
}
