% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{Easily describe datasets}
\usage{
crosstable(
  data,
  .vars = NULL,
  ...,
  by = NULL,
  total = c("none", "row", "column", "both"),
  margin = c("row", "column", "cell", "none", "all"),
  percent_digits = 2,
  showNA = c("ifany", "always", "no"),
  label = TRUE,
  funs = c(` ` = cross_summary),
  funs_arg = list(),
  cor_method = c("pearson", "kendall", "spearman"),
  test = FALSE,
  test_args = crosstable_test_args(),
  unique_numeric = 3,
  date_format = NULL,
  effect = FALSE,
  effect_args = crosstable_effect_args(),
  times = NULL,
  followup = FALSE
)
}
\arguments{
\item{data}{a data.frame}

\item{.vars}{the variables to describe. Can be a character or name vector, a tidyselect helper, a (lambda) function that returns a logical, or a formula. See examples or \code{vignette("crosstable-selection")} for more details.}

\item{...}{more variables to describe. Cannot be a lambda function nor a formula.}

\item{by}{the variable to group on. Character or name.}

\item{total}{one of ["none", "row", "column" or "both"] to indicate whether to add total rows and/or columns. Default to \code{none}.}

\item{margin}{one of ["row", "column", "cell", "none" or "all"] to indicate which proportions should be computed in frequency tables. Default to \code{row}.}

\item{percent_digits}{number of digits for percentages}

\item{showNA}{whether to show NA in categorial variables (one of \code{c("ifany", "always", "no")}, like in \code{table()})}

\item{label}{whether to show labels. See \link{import_labels} or \link{set_label}for how to add labels to the dataset columns.}

\item{funs}{functions to apply to numeric variables. Default to \link{cross_summary}.}

\item{funs_arg}{additionnal parameters for \code{funs}, e.g. \code{digits} (the number of decimal places) for the default \link{cross_summary}. Ultimately, these arguments are passed to \link{format_fixed}.}

\item{cor_method}{one of ["pearson", "kendall", or "spearman"] to indicate which correlation coefficient is to be used.}

\item{test}{whether to perform tests}

\item{test_args}{See \code{\link{crosstable_test_args}} to override default testing behaviour.}

\item{unique_numeric}{the number of non-missing different levels a variable should have to be considered as numeric}

\item{effect}{whether to compute a effect measure}

\item{effect_args}{See \code{\link{crosstable_effect_args}} to override default behaviour.}

\item{times}{when using formula with \code{\link[survival:Surv]{survival::Surv()}} objects, which times to summarize}

\item{followup}{when using formula with \code{\link[survival:Surv]{survival::Surv()}} objects, whether to display follow-up time}
}
\description{
Generate a descriptive table of all chosen columns, as contingency tables for categorical variables and as calculation summaries for numeric variables. If the \code{by} argument points to a categorical variable, \code{crosstable} will output a description of all columns for every level. Else, if it points to a numeric variable, \code{crosstable} will calculate correlation coefficients with all other selected numeric columns. Finally, if it points to a \code{Surv} object, \code{crosstable} will describe the survival at different times.
}
\details{
Can be formatted as an HTML table using \code{\link[=as_flextable]{as_flextable()}}.
}
\examples{
#whole table
crosstable(iris)
crosstable(mtcars)
crosstable(mtcars2)

#tidyselection, custom functions
library(dplyr)
crosstable(mtcars2, ends_with("t"), starts_with("c"), by=vs, 
           funs=c(mean, quantile), funs_arg = list(probs=c(.25,.75)))

#margin and totals
crosstable(mtcars2, disp, vs, by=am, 
           margin=c("row", "col"), total = "both")

#predicate selection, correlation, testing
crosstable(mtcars2, where(is.numeric), by=hp, test=TRUE)

#lambda selection, effect calculation
crosstable(mtcars2, ~is.numeric(.x) && mean(.x)>50, by=vs, effect=TRUE)

#Dates
mtcars2$my_date = as.Date(mtcars2$hp , origin="2010-01-01") \%>\% set_label("Some nonsense date")
crosstable(mtcars2, my_date, by=vs, date_format="\%d/\%m/\%Y")

#Survival data (using formula UI)
library(survival)
crosstable(aml, Surv(time, status) ~ x,times=c(0,15,30,150), followup=TRUE)
}
\seealso{
as_flextable, import_labels
}
