% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{crosstable_effect_args}
\alias{crosstable_effect_args}
\title{Default arguments for \code{\link[=crosstable]{crosstable()}} effects}
\usage{
crosstable_effect_args()
}
\value{
A list with testing parameters:
\itemize{
\item \code{effect.summarize} - a function of three arguments (continuous variable, grouping variable and conf.level), used to compare continuous variable. Returns a list of five components: \code{effect} (the effect value(s)), \code{ci} (the matrix of confidence interval(s)), \code{effect.name} (the interpretiation(s) of the effect value(s)), \code{effect.type} (the description of the measure used) and \code{conf.level} (the confidence interval level). See \code{diff_mean_auto}, \code{diff_mean_student} or \code{diff_mean_boot} for some examples of such functions. Users can provide their own function.
\item \code{effect.tabular} - a function of three arguments (two categorical variables and conf.level) used to measure the associations between two factors. Returns a list of five components: \code{effect} (the effect value(s)), \code{ci} (the matrix of confidence interval(s)), \code{effect.name} (the interpretation(s) of the effect value(s)), \code{effect.type} (the description of the measure used) and \code{conf.level} (the confidence interval level). See \code{effect_odd_ratio}, \code{effect_relative_risk}, \code{effect_risk_difference}, \code{effect_odd_ratio}, \code{rr.col.by.row}, or \code{rd.col.by.row} for some examples of such functions. Users can provide their own function.
\item \code{effect.survival} - a function of two argument (a formula and conf.level), used to measure the association between a consored and a factor. Returns the same components as created by \code{effect.summarize}. See \code{effect_survival_coxph}. Users can provide their own function.
\item \code{conf.level} - the desired confidence interval level
\item \code{digits} - the decimal places
\item \code{show.effect} - a function to format the effect. See \code{\link[=display_effect]{display_effect()}}.
}
}
\description{
Default arguments for \code{\link[=crosstable]{crosstable()}} effects
}
\author{
Dan Chaltiel
}
