% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable-package.R
\name{crosstable_options}
\alias{crosstable_options}
\title{Options for the package \code{crosstable}}
\description{
Here is a comprehensive list of all options that you can set globally.

For flextables:
\itemize{
\item \code{crosstable_autofit}:
\item \code{crosstable_compact}:
\item \code{crosstable_show_test_name}:
\item \code{crosstable_compact_padding}:
\item \code{crosstable_wrap_id}: if id contains no spaces, wrap it with this maximum number of characters
}

For rounding:
\itemize{
\item \code{crosstable_only_round}:
}

For specifying styles in officer docx output:
\itemize{
\item \code{crosstable_style_normal}:
\item \code{crosstable_style_heading}:
\item \code{crosstable_style_strong}:
\item \code{crosstable_style_image}:
\item \code{crosstable_style_legend}:
\item \code{crosstable_style_list_ordered}:
\item \code{crosstable_style_list_unordered}:
}
}
