% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{crosstable_reset_options}
\alias{crosstable_reset_options}
\title{Reset all \code{crosstable} options.}
\usage{
crosstable_reset_options(verbose = TRUE)
}
\arguments{
\item{keep_null}{set to TRUE to get a list}
}
\value{
Nothing, called for its side effects
}
\description{
Reset all \code{crosstable} options.
}
