% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{crosstable_test_args}
\alias{crosstable_test_args}
\title{Default arguments for \code{\link{crosstable}} tests}
\usage{
crosstable_test_args()
}
\value{
A list with testing parameters:
\itemize{
\item \code{test.summarize} - a function of two arguments (continuous variable and grouping variable), used to compare continuous variable. Returns a list of two components : \code{p.value} and \code{method} (the test name). See \link{test.summarize.auto}, \link{test.tabular.auto}, \link{test.survival.logrank}, or \link{test.summarize.linear.contrasts} for some examples of such functions. Users can provide their own function.
\item \code{test.survival} - a function of one argument (a formula), used to compare survival estimations. Returns the same components as created by \code{test.summarize}. See \code{\link{test.survival.logrank}}. Users can provide their own function.
\item \code{test.tabular} - a function of two arguments (two categorical variables), used to test association between two factors.  Returns the same components as created by \code{test.summarize}. See \code{\link{test.tabular.auto}} and \code{\link{test.tabular.fisher}}. Users can provide their own function.
\item \code{show.test} - function used to display the test result. See \code{\link{display.test}}.
\item \code{plim} - number of digits for the p value
\item \code{show.method} - whether to display the test name (logical)
}
}
\description{
Default arguments for \code{\link{crosstable}} tests
}
\author{
Dan Chaltiel
}
