% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{display_effect}
\alias{display_effect}
\title{Default function to display the effect}
\usage{
display_effect(effect, digits = 4)
}
\arguments{
\item{effect}{effect}

\item{digits}{digits}
}
\description{
Default function to display the effect
}
