% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{effect.survival.coxph}
\alias{effect.survival.coxph}
\title{Effect measure for association between one consored variable and one categorical variable}
\usage{
effect.survival.coxph(formula, conf.level = 0.95)
}
\arguments{
\item{formula}{a formula}
}
\value{
a list with two componments: p.value and method
}
\description{
Effect measure for association between one consored variable and one categorical variable
}
\author{
David Hajage
}
