% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_summary}
\alias{effect_summary}
\alias{diff_mean_auto}
\alias{diff_mean_boot}
\alias{diff_median}
\alias{diff_mean_student}
\title{Effect measure for association between one continuous and one categorical variable}
\usage{
diff_mean_auto(x, g, conf_level = 0.95, R = 500)

diff_mean_boot(x, g, conf_level = 0.95, R = 500)

diff_median(x, g, conf_level = 0.95, R = 500)

diff_mean_student(x, g, conf_level = 0.95)
}
\arguments{
\item{x}{vector (of exactly 2 unique levels)}

\item{g}{another vector}

\item{conf_level}{confidence interval level}

\item{R}{number of bootstrap replication}
}
\value{
A list with five componments: effect, ci, effect.name, effect.type, and conf_level
}
\description{
User can either use or extend these functions to parametrize effect calculation.
}
\section{Functions}{
\itemize{
\item \code{diff_mean_auto}: (\strong{Default}) calculate a specific "difference in means" effect based on normality (Shapiro or Anderson test) and variance homogeneity (Bartlett test)

\item \code{diff_mean_boot}: calculate a "difference in means" effect by bootstraping

\item \code{diff_median}: calculate a "difference in medians" effect by bootstraping

\item \code{diff_mean_student}: calculate a "difference in means" effect using \code{t.test} confidence intervals
}}

\seealso{
\code{\link[=crosstable_effect_args]{crosstable_effect_args()}}
}
