% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_survival}
\alias{effect_survival}
\alias{effect_survival_coxph}
\title{Effect measure for association between one censored variable and one categorical variable}
\usage{
effect_survival_coxph(x, by, conf_level = 0.95)
}
\arguments{
\item{x}{survival vector (made using \code{\link[survival:Surv]{survival::Surv()}})}

\item{by}{categorical vector (of exactly 2 unique levels)}

\item{conf_level}{confidence interval level}
}
\value{
a list with two components: p.value and method
}
\description{
Effect measure for association between one censored variable and one categorical variable
}
\author{
Dan Chaltiel, David Hajage
}
