% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{expand_formula}
\alias{expand_formula}
\title{Expand ... in a formula}
\usage{
expand_formula(formula, varnames)
}
\arguments{
\item{formula}{formula (character)}

\item{varnames}{variables names}
}
\description{
Expand ... in a formula
}
\author{
David Hajage
}
\keyword{internal}
