% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.r
\name{format_fixed}
\alias{format_fixed}
\title{Format numbers with same number of decimals, including trailing zeros}
\usage{
format_fixed(x, digits, zero_digits = 1, only_round = FALSE)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimals}

\item{zero_digits}{number of significant digits for values rounded to 0 (set to NULL to keep the original 0 value)}

\item{only_round}{util option, to simply return rounded value instead of formatted}
}
\value{
a character vector of formatted numbers
}
\description{
Format numbers with same number of decimals, including trailing zeros
}
\examples{
x = c(1, 1.2, 12.78749, pi, 0.00000012)
format_fixed(x, digits=3)
format_fixed(x, digits=3, zero_digits=2)
format_fixed(x, digits=3, zero_digits=NULL)
x = iris$Sepal.Length/10000
x \%>\% 
sd(na.rm=na.rm) \%>\% 
  format_fixed(dig=3, zero_digits=2, only_round=T)
}
\author{
Dan Chaltiel
}
