% Generated by roxygen2 (4.0.2): do not edit by hand
\name{freq.data.frame}
\alias{freq.data.frame}
\title{Compute frequencies (data.frame input)}
\usage{
freq.data.frame(df, showNA = c("no", "ifany", "always"), total = FALSE,
  digits = 2, label = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{showNA}{showNA}
}
\description{
Compute frequencies (data.frame input)
}
\author{
David Hajage
}
\keyword{internal}

