% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.r
\name{freq}
\alias{freq}
\title{Compute frequencies}
\usage{
freq(x, showNA = c("no", "ifany", "always"), total = FALSE, digits = 2)
}
\arguments{
\item{x}{factor}

\item{showNA}{showNA}

\item{total}{total}

\item{digits}{digits}
}
\description{
Compute frequencies
}
\author{
David Hajage
}
\keyword{internal}
