% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_label}
\alias{get_label}
\title{Get label if wanted and available, or default (name) otherwise}
\usage{
get_label(x, default = names(x))
}
\arguments{
\item{x}{labelled object}

\item{default}{value returned if there is no label. Default to \code{names(x)}.}
}
\description{
Get label if wanted and available, or default (name) otherwise
}
\examples{
#vectors
get_label(mtcars2$mpg)
get_label(mtcars$mpg)
get_label(mtcars$mpg, default="foo")
get_label(list(bar=mtcars$mpg)) #default to names

#data.frames
get_label(mtcars2["mpg"])
get_label(mtcars["mpg"]) #default to names
get_label(mtcars["mpg"], default="bar")
}
\seealso{
\link{set_label}, \link{import_labels}, \link{remove_label}, \link[Hmisc:label]{Hmisc::label}, \link[expss:var_lab]{expss::var_lab}
}
