% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_categorical.R
\name{get_percent_pattern}
\alias{get_percent_pattern}
\title{Percent pattern helper}
\usage{
get_percent_pattern(
  margin = c("row", "column", "cell", "none", "all"),
  na = FALSE
)
}
\arguments{
\item{margin}{a vector giving the margins to compute.}

\item{na}{whether to use \code{NA}}
}
\value{
a list
}
\description{
Get a list with pre-filled values for \code{percent_pattern}.
}
\examples{
get_percent_pattern(c("cells","row","column"))
get_percent_pattern(c("cells","row","column"), na=TRUE)
}
