% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.crosstable}
\alias{is.crosstable}
\alias{is.transposed_crosstable}
\alias{is.compacted_crosstable}
\alias{is.multiby_crosstable}
\title{Test if an object is a crosstable}
\usage{
is.crosstable(x)

is.transposed_crosstable(x)

is.compacted_crosstable(x)

is.multiby_crosstable(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the \code{crosstable} class or other subclasses.
}
\description{
Test if an object is a crosstable
}
