% Generated by roxygen2 (4.0.2): do not edit by hand
\name{left_right}
\alias{left_right}
\title{Separate left and right part of a formula}
\usage{
left_right(formula)
}
\arguments{
\item{formula}{formula (character)}
}
\description{
Separate left and right part of a formula
}
\author{
David Hajage
}
\keyword{internal}

