% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.r
\name{Max}
\alias{Max}
\title{Return the max (and have formals)}
\usage{
Max(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{Remove NA?}
}
\description{
Return the max (and have formals)
}
\author{
David Hajage
}
\keyword{univar}

