% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.r
\name{minmax}
\alias{minmax}
\title{Return min and max}
\usage{
minmax(x, na.rm = TRUE, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{\code{TRUE} as default}

\item{dig}{number of digits}
}
\description{
Return min and max
}
\keyword{internal}

