% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funs.r
\name{mysummary}
\alias{mysummary}
\title{Summarize a numeric vector}
\usage{
mysummary(x, na.rm = TRUE, dig = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{\code{TRUE} as default}

\item{dig}{number of digits}
}
\description{
Summarize a numeric vector
}
\keyword{internal}

