% Generated by roxygen2 (4.0.1): do not edit by hand
\name{n}
\alias{n}
\title{Return the number of non NA observations}
\usage{
n(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{not used}
}
\description{
Return the number of non NA observations
}
\author{
David Hajage
}
\keyword{univar}

