% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{na}
\alias{na}
\title{Return the number of NA observations}
\usage{
na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
integer, number of NA observations
}
\description{
Return the number of NA observations
}
\author{
David Hajage
}
