% Generated by roxygen2 (4.0.2): do not edit by hand
\name{na}
\alias{na}
\title{Return the number of NA observations}
\usage{
na(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{not used}
}
\description{
Return the number of NA observations
}
\author{
David Hajage
}
\keyword{univar}

