% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula}
\usage{
parse_formula(formula, varnames)
}
\arguments{
\item{formula}{formula (character)}

\item{varnames}{variables names}
}
\description{
Parse a formula
}
\author{
David Hajage
}
\keyword{internal}
