% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plim}
\alias{plim}
\title{Format p values (alternative to \code{\link[=format.pval]{format.pval()}})}
\usage{
plim(p, digits = 4)
}
\arguments{
\item{p}{p values}

\item{digits}{number of digits}
}
\value{
formatted p values
}
\description{
Format p values (alternative to \code{\link[=format.pval]{format.pval()}})
}
\seealso{
\code{\link[=format.pval]{format.pval()}}, https://stackoverflow.com/a/23018806/3888000
}
\author{
David Hajage
}
