% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regroup}
\alias{regroup}
\title{Regroup factors with factors, and numerical variables with numerical variables}
\usage{
regroup(vars, numdata, catdata, survdata)
}
\arguments{
\item{vars}{vars}

\item{numdata}{numdata}

\item{catdata}{catdata}

\item{survdata}{survdata}
}
\description{
Regroup factors with factors, and numerical variables with numerical variables
}
\author{
David Hajage
}
\keyword{internal}

