% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_dataframe_with_labels}
\alias{rename_dataframe_with_labels}
\title{Rename every column of a dataframe with its label}
\usage{
rename_dataframe_with_labels(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
A dataframe, as \code{df}, which names are copied from the label attribute
}
\description{
Rename every column of a dataframe with its label
}
\examples{
library(dplyr)
mtcars2 \%>\% 
  select(1:5) \%>\% 
  rename_dataframe_with_labels()
}
\author{
Dan Chaltiel
}
