% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_with_labels}
\alias{rename_with_labels}
\alias{rename_dataframe_with_labels}
\title{Rename every column of a dataframe with its label}
\usage{
rename_with_labels(df, cols = everything(), except = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{cols}{<\code{\link[tidyselect:language]{tidy-select}}> columns to be renamed.}

\item{except}{<\code{\link[tidyselect:language]{tidy-select}}> columns not be renamed. Deprecated in favor of \code{cols}.}
}
\value{
A dataframe which names are copied from the label attribute
}
\description{
Rename every column of a dataframe with its label
}
\examples{
rename_with_labels(mtcars2[,1:5], cols=1:4) \%>\% names()
rename_with_labels(iris2, cols=-c(Sepal.Length, Sepal.Width)) \%>\% names()
}
