% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{set_label}
\alias{set_label}
\title{Set the "label" attribute of an object}
\usage{
set_label(x, value)
}
\arguments{
\item{x}{object to labelise}

\item{value}{value of the label}
}
\description{
Set the "label" attribute of an object
}
\examples{
library(dplyr)
mtcars \%>\% 
   mutate(mpg2=set_label(mpg, "Miles per gallon")) \%>\% 
   crosstable(mpg, mpg2)
}
\seealso{
\link{get_label}, \link{import_labels}, \link{remove_label}
}
