% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/survival.r
\name{survival.data.frame.by}
\alias{survival.data.frame.by}
\title{Compute survival according to a factor (data.frame input)}
\usage{
survival.data.frame.by(df, by, times = NULL, followup = FALSE,
  total = FALSE, digits = 2, test = FALSE,
  test.survival = test.survival.logrank, show.test = display.test,
  plim = 4, show.method = TRUE, label = FALSE)
}
\arguments{
\item{df}{df}

\item{by}{by}

\item{times}{times}

\item{followup}{followup}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.survival}{test.survival}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}

\item{label}{label}
}
\description{
Compute survival according to a factor (data.frame input)
}
\author{
David Hajage
}

