% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.r
\name{survival.data.frame}
\alias{survival.data.frame}
\title{Compute survival (data.frame input)}
\usage{
survival.data.frame(df, times = NULL, digits = 2, followup = FALSE,
  label = FALSE)
}
\arguments{
\item{df}{df}

\item{times}{times}

\item{digits}{digits}

\item{followup}{followup}

\item{label}{label}
}
\description{
Compute survival (data.frame input)
}
\author{
David Hajage
}
