% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabular}
\alias{tabular}
\title{Compute a contingency table}
\usage{
tabular(x, y, showNA = c("no", "ifany", "always"), margin = 0:2,
  total = FALSE, digits = 2, test = FALSE,
  test.tabular = test.tabular.auto, show.test = display.test, plim = 4,
  show.method = TRUE)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{showNA}{showNA}

\item{margin}{margin}

\item{total}{total}

\item{digits}{digits}

\item{test}{test}

\item{test.tabular}{test.tabular}

\item{show.test}{show.test}

\item{plim}{plim}

\item{show.method}{show.method}
}
\description{
Compute a contingency table
}
\author{
David Hajage
}
\keyword{internal}

