% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test.summarize.contrasts.lin}
\alias{test.summarize.contrasts.lin}
\title{Test for linear trend across ordered factor with contrasts}
\usage{
test.summarize.contrasts.lin(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{ordered factor}
}
\value{
a list with two componments: p.value and method
}
\description{
Test for linear trend across ordered factor with contrasts
}
\examples{
library(dplyr)
iris \%>\% 
  mutate(Petal.Width.qt = paste0("Q", ntile(Petal.Width, 5)) \%>\% ordered()) \%>\% 
  cross(Petal.Length ~ Petal.Width.qt, data=., test=T, test.summarize = test.summarize.contrasts.lin)
}
\author{
Dan Chaltiel
}
