% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test.tabular.fisher}
\alias{test.tabular.fisher}
\title{test for contingency table}
\usage{
test.tabular.fisher(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}
}
\value{
a list with two componments: p.value and method
}
\description{
Compute a fisher test
}
\author{
David Hajage
}
